Network.AWS.DataPipeline

Description

AWS Data Pipeline is a web service that you can use to automate the movement and transformation of data. With AWS Data Pipeline, you can define data-driven workflows, so that tasks can be dependent on the successful completion of previous tasks.

Documentation

module Network.AWS.DataPipeline.ActivatePipeline

module Network.AWS.DataPipeline.AddTags

module Network.AWS.DataPipeline.CreatePipeline

module Network.AWS.DataPipeline.DeactivatePipeline

module Network.AWS.DataPipeline.DeletePipeline

module Network.AWS.DataPipeline.DescribeObjects

module Network.AWS.DataPipeline.DescribePipelines

module Network.AWS.DataPipeline.EvaluateExpression

module Network.AWS.DataPipeline.GetPipelineDefinition

module Network.AWS.DataPipeline.ListPipelines

module Network.AWS.DataPipeline.PollForTask

module Network.AWS.DataPipeline.PutPipelineDefinition

module Network.AWS.DataPipeline.QueryObjects

module Network.AWS.DataPipeline.RemoveTags

module Network.AWS.DataPipeline.ReportTaskProgress

module Network.AWS.DataPipeline.ReportTaskRunnerHeartbeat

module Network.AWS.DataPipeline.SetStatus

module Network.AWS.DataPipeline.SetTaskStatus

module Network.AWS.DataPipeline.Types

module Network.AWS.DataPipeline.ValidatePipelineDefinition