Network.AWS.DataPipeline.ActivatePipeline

Description

Validates a pipeline and initiates processing. If the pipeline does not pass validation, activation fails. You cannot perform this operation on FINISHED pipelines and attempting to do so will return an InvalidRequestException.

Call this action to start processing pipeline tasks of a pipeline you've created using the CreatePipeline and PutPipelineDefinition actions. A pipeline cannot be modified after it has been successfully activated.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_ActivatePipeline.html

Request

data ActivatePipeline

Request constructor

activatePipeline

Arguments

:: Text
apPipelineId

-> ActivatePipeline

ActivatePipeline constructor.

The fields accessible through corresponding lenses are:

apParameterValues :: [ParameterValue]

apPipelineId :: Text

Request lenses

apParameterValues :: Lens' ActivatePipeline [ParameterValue]

Returns a list of parameter values to pass to the pipeline at activation.

apPipelineId :: Lens' ActivatePipeline Text

The identifier of the pipeline to activate.

Response

data ActivatePipelineResponse

Response constructor

activatePipelineResponse :: ActivatePipelineResponse

ActivatePipelineResponse constructor.