Network.AWS.DataPipeline.CreatePipeline

Description

Creates a new empty pipeline. When this action succeeds, you can then use the PutPipelineDefinition action to populate the pipeline.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_CreatePipeline.html

Request

data CreatePipeline

Request constructor

createPipeline

Arguments

:: Text
cpName

-> Text
cpUniqueId

-> CreatePipeline

CreatePipeline constructor.

The fields accessible through corresponding lenses are:

cpDescription :: Maybe Text

cpName :: Text

cpUniqueId :: Text

Request lenses

cpDescription :: Lens' CreatePipeline (Maybe Text)

The description of the new pipeline.

cpName :: Lens' CreatePipeline Text

The name of the new pipeline. You can use the same name for multiple pipelines associated with your AWS account, because AWS Data Pipeline assigns each new pipeline a unique pipeline identifier.

cpUniqueId :: Lens' CreatePipeline Text

A unique identifier that you specify. This identifier is not the same as the pipeline identifier assigned by AWS Data Pipeline. You are responsible for defining the format and ensuring the uniqueness of this identifier. You use this parameter to ensure idempotency during repeated calls to CreatePipeline. For example, if the first call to CreatePipeline does not return a clear success, you can pass in the same unique identifier and pipeline name combination on a subsequent call to CreatePipeline. CreatePipeline ensures that if a pipeline already exists with the same name and unique identifier, a new pipeline will not be created. Instead, you'll receive the pipeline identifier from the previous attempt. The uniqueness of the name and unique identifier combination is scoped to the AWS account or IAM user credentials.

Response

data CreatePipelineResponse

Response constructor

createPipelineResponse

Arguments

:: Text
cprPipelineId

-> CreatePipelineResponse

CreatePipelineResponse constructor.

The fields accessible through corresponding lenses are:

cprPipelineId :: Text

Response lenses

cprPipelineId :: Lens' CreatePipelineResponse Text

The ID that AWS Data Pipeline assigns the newly created pipeline. The ID is a string of the form: df-06372391ZG65EXAMPLE.