Network.AWS.DataPipeline.DeactivatePipeline

Description

Deactivates the specified running pipeline. The pipeline is set to the DEACTIVATING state until the deactivation process completes.

To resume a deactivated pipeline, use ActivatePipeline. By default, the pipeline resumes from the last completed execution. Optionally, you can specify the date and time to resume the pipeline.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_DeactivatePipeline.html

Request

data DeactivatePipeline

Request constructor

deactivatePipeline

Arguments

:: Text
dp1PipelineId

-> DeactivatePipeline

DeactivatePipeline constructor.

The fields accessible through corresponding lenses are:

dp1CancelActive :: Maybe Bool

dp1PipelineId :: Text

Request lenses

dp1CancelActive :: Lens' DeactivatePipeline (Maybe Bool)

Indicates whether to cancel any running objects. The default is true, which sets the state of any running objects to CANCELED. If this value is false, the pipeline is deactivated after all running objects finish.

dp1PipelineId :: Lens' DeactivatePipeline Text

The ID of the pipeline.

Response

data DeactivatePipelineResponse

Response constructor

deactivatePipelineResponse :: DeactivatePipelineResponse

DeactivatePipelineResponse constructor.