Network.AWS.DataPipeline.DeletePipeline

Description

Permanently deletes a pipeline, its pipeline definition and its run history. You cannot query or restore a deleted pipeline. AWS Data Pipeline will attempt to cancel instances associated with the pipeline that are currently being processed by task runners. Deleting a pipeline cannot be undone. To temporarily pause a pipeline instead of deleting it, call SetStatus with the status set to Pause on individual components. Components that are paused by SetStatus can be resumed. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.DeletePipeline Content-Length: 50 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE"} x-amzn-RequestId: b7a88c81-0754-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 0 Date: Mon, 12 Nov 2012 17:50:53 GMT Unexpected response: 200, OK, undefined.

Request

data DeletePipeline

The input for the DeletePipeline action.

Request constructor

deletePipeline

Arguments

:: Text
dpPipelineId

-> DeletePipeline

Smart constructor for the minimum required parameters to construct a valid DeletePipeline request.

The fields accessible through corresponding lenses are:

PipelineId :: Text

Request lenses

dpPipelineId :: Lens' DeletePipeline Text

The identifier of the pipeline to be deleted.

Response

data DeletePipelineResponse

Response constructor

deletePipelineResponse :: DeletePipelineResponse

Smart constructor for the minimum required parameters to construct a valid DeletePipelineResponse response.

This constructor is provided for convenience and testing purposes.