Network.AWS.DataPipeline.EvaluateExpression

Description

Evaluates a string in the context of a specified object. A task runner can use this action to evaluate SQL queries stored in Amazon S3.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_EvaluateExpression.html

Request

data EvaluateExpression

Request constructor

evaluateExpression

Arguments

:: Text
eePipelineId

-> Text
eeObjectId

-> Text
eeExpression

-> EvaluateExpression

EvaluateExpression constructor.

The fields accessible through corresponding lenses are:

eeExpression :: Text

eeObjectId :: Text

eePipelineId :: Text

Request lenses

eeExpression :: Lens' EvaluateExpression Text

The expression to evaluate.

eeObjectId :: Lens' EvaluateExpression Text

The identifier of the object.

eePipelineId :: Lens' EvaluateExpression Text

The identifier of the pipeline.

Response

data EvaluateExpressionResponse

Response constructor

evaluateExpressionResponse

Arguments

:: Text
eerEvaluatedExpression

-> EvaluateExpressionResponse

EvaluateExpressionResponse constructor.

The fields accessible through corresponding lenses are:

eerEvaluatedExpression :: Text

Response lenses

eerEvaluatedExpression :: Lens' EvaluateExpressionResponse Text

The evaluated expression.