Network.AWS.DataPipeline.GetPipelineDefinition

Description

Returns the definition of the specified pipeline. You can call GetPipelineDefinition to retrieve the pipeline definition you provided using PutPipelineDefinition.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_GetPipelineDefinition.html

Request

data GetPipelineDefinition

Request constructor

getPipelineDefinition

Arguments

:: Text
gpdPipelineId

-> GetPipelineDefinition

GetPipelineDefinition constructor.

The fields accessible through corresponding lenses are:

gpdPipelineId :: Text

gpdVersion :: Maybe Text

Request lenses

gpdPipelineId :: Lens' GetPipelineDefinition Text

The identifier of the pipeline.

gpdVersion :: Lens' GetPipelineDefinition (Maybe Text)

The version of the pipeline definition to retrieve. This parameter accepts the values latest (default) and active. Where latest indicates the last definition saved to the pipeline and active indicates the last definition of the pipeline that was activated.

Response

data GetPipelineDefinitionResponse

Response constructor

getPipelineDefinitionResponse :: GetPipelineDefinitionResponse

GetPipelineDefinitionResponse constructor.

The fields accessible through corresponding lenses are:

gpdrPipelineObjects :: [PipelineObject]

Response lenses

gpdrPipelineObjects :: Lens' GetPipelineDefinitionResponse [PipelineObject]

An array of objects defined in the pipeline.