Network.AWS.DataPipeline.ListPipelines

Description

Returns a list of pipeline identifiers for all active pipelines. Identifiers are returned only for pipelines you have permission to access. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.ListPipelines Content-Length: 14 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {} Status: x-amzn-RequestId: b3104dc5-0734-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 39 Date: Mon, 12 Nov 2012 17:50:53 GMT {PipelineIdList: [ {"id": "df-08785951KAKJEXAMPLE", "name": MyPipeline}, {"id": "df-08662578ISYEXAMPLE", "name": MySecondPipeline} ] }.

Request

data ListPipelines

The input to the ListPipelines action.

Request constructor

listPipelines :: ListPipelines

Smart constructor for the minimum required parameters to construct a valid ListPipelines request.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

Request lenses

lpMarker :: Lens' ListPipelines (Maybe Text)

The starting point for the results to be returned. The first time you call ListPipelines, this value should be empty. As long as the action returns HasMoreResults as True, you can call ListPipelines again and pass the marker value from the response to retrieve the next set of results.

Response

data ListPipelinesResponse

Contains the output from the ListPipelines action.

Response constructor

listPipelinesResponse

Arguments

:: [PipelineIdName]
lprPipelineIdList

-> Bool
lprHasMoreResults

-> ListPipelinesResponse

Smart constructor for the minimum required parameters to construct a valid ListPipelinesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

PipelineIdList :: [PipelineIdName]

Marker :: Maybe Text

HasMoreResults :: Bool

Response lenses

lprPipelineIdList :: Lens' ListPipelinesResponse [PipelineIdName]

A list of all the pipeline identifiers that your account has permission to access. If you require additional information about the pipelines, you can use these identifiers to call DescribePipelines and GetPipelineDefinition.

lprMarker :: Lens' ListPipelinesResponse (Maybe Text)

If not null, indicates the starting point for the set of pipeline identifiers that the next call to ListPipelines will retrieve. If null, there are no more pipeline identifiers.

lprHasMoreResults :: Lens' ListPipelinesResponse Bool

If True, there are more results that can be obtained by a subsequent call to ListPipelines.