Network.AWS.DataPipeline.QueryObjects

Description

Queries a pipeline for the names of objects that match a specified set of conditions.

The objects returned by QueryObjects are paginated and then filtered by the value you set for query. This means the action may return an empty result set with a value set for marker. If HasMoreResults is set to True, you should continue to call QueryObjects, passing in the returned value for marker, until HasMoreResults returns False.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_QueryObjects.html

Request

data QueryObjects

Request constructor

queryObjects

Arguments

:: Text
qoPipelineId

-> Text
qoSphere

-> QueryObjects

QueryObjects constructor.

The fields accessible through corresponding lenses are:

qoLimit :: Maybe Int

qoMarker :: Maybe Text

qoPipelineId :: Text

qoQuery :: Maybe Query

qoSphere :: Text

Request lenses

qoLimit :: Lens' QueryObjects (Maybe Int)

Specifies the maximum number of object names that QueryObjects will return in a single call. The default value is 100.

qoMarker :: Lens' QueryObjects (Maybe Text)

The starting point for the results to be returned. The first time you call QueryObjects, this value should be empty. As long as the action returns HasMoreResults as True, you can call QueryObjects again and pass the marker value from the response to retrieve the next set of results.

qoPipelineId :: Lens' QueryObjects Text

Identifier of the pipeline to be queried for object names.

qoQuery :: Lens' QueryObjects (Maybe Query)

Query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

qoSphere :: Lens' QueryObjects Text

Specifies whether the query applies to components or instances. Allowable values: COMPONENT, INSTANCE, ATTEMPT.

Response

data QueryObjectsResponse

Response constructor

queryObjectsResponse :: QueryObjectsResponse

QueryObjectsResponse constructor.

The fields accessible through corresponding lenses are:

qorHasMoreResults :: Maybe Bool

qorIds :: [Text]

qorMarker :: Maybe Text

Response lenses

qorHasMoreResults :: Lens' QueryObjectsResponse (Maybe Bool)

If True, there are more results that can be obtained by a subsequent call to QueryObjects.

qorIds :: Lens' QueryObjectsResponse [Text]

A list of identifiers that match the query selectors.

qorMarker :: Lens' QueryObjectsResponse (Maybe Text)

The starting point for the results to be returned. As long as the action returns HasMoreResults as True, you can call QueryObjects again and pass the marker value from the response to retrieve the next set of results.