Network.AWS.DataPipeline.ReportTaskProgress

Description

Updates the AWS Data Pipeline service on the progress of the calling task runner. When the task runner is assigned a task, it should call ReportTaskProgress to acknowledge that it has the task within 2 minutes. If the web service does not recieve this acknowledgement within the 2 minute window, it will assign the task in a subsequent PollForTask> call. After this initial acknowledgement, the task runner only needs to report progress every 15 minutes to maintain its ownership of the task. You can change this reporting time from 15 minutes by specifying a reportProgressTimeout field in your pipeline. If a task runner does not report its status after 5 minutes, AWS Data Pipeline will assume that the task runner is unable to process the task and will reassign the task in a subsequent response to PollForTask>. task runners should call ReportTaskProgress> every 60 seconds.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_ReportTaskProgress.html

Request

data ReportTaskProgress

Request constructor

reportTaskProgress

Arguments

:: Text
rtpTaskId

-> ReportTaskProgress

ReportTaskProgress constructor.

The fields accessible through corresponding lenses are:

rtpTaskId :: Text

Request lenses

rtpTaskId :: Lens' ReportTaskProgress Text

Identifier of the task assigned to the task runner. This value is provided in the TaskObject> that the service returns with the response for the PollForTask> action.

Response

data ReportTaskProgressResponse

Response constructor

reportTaskProgressResponse

Arguments

:: Bool
rtprCanceled

-> ReportTaskProgressResponse

ReportTaskProgressResponse constructor.

The fields accessible through corresponding lenses are:

rtprCanceled :: Bool

Response lenses

rtprCanceled :: Lens' ReportTaskProgressResponse Bool

If True, the calling task runner should cancel processing of the task. The task runner does not need to call SetTaskStatus> for canceled tasks.