Network.AWS.DataPipeline.ReportTaskRunnerHeartbeat

Description

Task runners call ReportTaskRunnerHeartbeat every 15 minutes to indicate that they are operational. If the AWS Data Pipeline Task Runner is launched on a resource managed by AWS Data Pipeline, the web service can use this call to detect when the task runner application has failed and restart a new instance.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_ReportTaskRunnerHeartbeat.html

Request

data ReportTaskRunnerHeartbeat

Request constructor

reportTaskRunnerHeartbeat

Arguments

:: Text
rtrhTaskrunnerId

-> ReportTaskRunnerHeartbeat

ReportTaskRunnerHeartbeat constructor.

The fields accessible through corresponding lenses are:

rtrhHostname :: Maybe Text

rtrhTaskrunnerId :: Text

rtrhWorkerGroup :: Maybe Text

Request lenses

rtrhHostname :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

The public DNS name of the task runner.

rtrhTaskrunnerId :: Lens' ReportTaskRunnerHeartbeat Text

The ID of the task runner. This value should be unique across your AWS account. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service provides a unique identifier when it launches the application. If you have written a custom task runner, you should assign a unique identifier for the task runner.

rtrhWorkerGroup :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

Response

data ReportTaskRunnerHeartbeatResponse

Response constructor

reportTaskRunnerHeartbeatResponse

Arguments

:: Bool
rtrhrTerminate

-> ReportTaskRunnerHeartbeatResponse

ReportTaskRunnerHeartbeatResponse constructor.

The fields accessible through corresponding lenses are:

rtrhrTerminate :: Bool

Response lenses

rtrhrTerminate :: Lens' ReportTaskRunnerHeartbeatResponse Bool

Indicates whether the calling task runner should terminate.