Network.AWS.DataPipeline.SetStatus

Description

Requests that the status of an array of physical or logical pipeline objects be updated in the pipeline. This update may not occur immediately, but is eventually consistent. The status that can be set depends on the type of object.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_SetStatus.html

Request

data SetStatus

Request constructor

setStatus

Arguments

:: Text
ssPipelineId

-> Text
ssStatus

-> SetStatus

SetStatus constructor.

The fields accessible through corresponding lenses are:

ssObjectIds :: [Text]

ssPipelineId :: Text

ssStatus :: Text

Request lenses

ssObjectIds :: Lens' SetStatus [Text]

Identifies an array of objects. The corresponding objects can be either physical or components, but not a mix of both types.

ssPipelineId :: Lens' SetStatus Text

Identifies the pipeline that contains the objects.

ssStatus :: Lens' SetStatus Text

Specifies the status to be set on all the objects in objectIds. For components, this can be either PAUSE or RESUME. For instances, this can be either CANCEL, RERUN, or MARK_FINISHED.

Response

data SetStatusResponse

Response constructor

setStatusResponse :: SetStatusResponse

SetStatusResponse constructor.