Network.AWS.DataPipeline.V2012_10_29

Description

AWS Data Pipeline is a web service that you can use to automate the movement and transformation of data. With AWS Data Pipeline, you can define data-driven workflows, so that tasks can be dependent on the successful completion of previous tasks.

Documentation

module Network.AWS.DataPipeline.V2012_10_29.ActivatePipeline

module Network.AWS.DataPipeline.V2012_10_29.CreatePipeline

module Network.AWS.DataPipeline.V2012_10_29.DeletePipeline

module Network.AWS.DataPipeline.V2012_10_29.DescribeObjects

module Network.AWS.DataPipeline.V2012_10_29.DescribePipelines

module Network.AWS.DataPipeline.V2012_10_29.EvaluateExpression

module Network.AWS.DataPipeline.V2012_10_29.GetPipelineDefinition

module Network.AWS.DataPipeline.V2012_10_29.ListPipelines

module Network.AWS.DataPipeline.V2012_10_29.PollForTask

module Network.AWS.DataPipeline.V2012_10_29.PutPipelineDefinition

module Network.AWS.DataPipeline.V2012_10_29.QueryObjects

module Network.AWS.DataPipeline.V2012_10_29.ReportTaskProgress

module Network.AWS.DataPipeline.V2012_10_29.ReportTaskRunnerHeartbeat

module Network.AWS.DataPipeline.V2012_10_29.SetStatus

module Network.AWS.DataPipeline.V2012_10_29.SetTaskStatus

module Network.AWS.DataPipeline.V2012_10_29.Types

module Network.AWS.DataPipeline.V2012_10_29.ValidatePipelineDefinition