Network.AWS.DataPipeline.V2012_10_29.EvaluateExpression

Description

Evaluates a string in the context of a specified object. A task runner can use this action to evaluate SQL queries stored in Amazon S3. POST / HTTP1.1 Content-Type: applicationx-amz-json-1.1 X-Amz-Target: DataPipeline.DescribePipelines Content-Length: 164 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-08785951KAKJEXAMPLE", "objectId": Schedule, "expression": "Transform started at #{startDateTime} and finished at #{endDateTime}"} x-amzn-RequestId: 02870eb7-0736-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 103 Date: Mon, 12 Nov 2012 17:50:53 GMT {"evaluatedExpression": "Transform started at 2012-12-12T00:00:00 and finished at 2012-12-21T18:00:00"}.

Documentation

data EvaluateExpression

Constructors

EvaluateExpression

Fields

_eeiPipelineId :: Text

The identifier of the pipeline.

_eeiObjectId :: Text

The identifier of the object.

_eeiExpression :: Text

The expression to evaluate.

eeiPipelineId :: Lens' EvaluateExpression Text

eeiObjectId :: Lens' EvaluateExpression Text

eeiExpression :: Lens' EvaluateExpression Text

data EvaluateExpressionResponse

Constructors

EvaluateExpressionResponse

Fields

_eeoEvaluatedExpression :: Text

The evaluated expression.

eeoEvaluatedExpression :: Iso' EvaluateExpressionResponse Text