Network.AWS.DataPipeline.V2012_10_29.Monadic

Description

AWS Data Pipeline is a web service that you can use to automate the movement and transformation of data. With AWS Data Pipeline, you can define data-driven workflows, so that tasks can be dependent on the successful completion of previous tasks.

The State operator variants from Control.Lens.Setter such as .= can be used to modify any additional request parameters before sending.

ActivatePipeline

Validates a pipeline and initiates processing. If the pipeline does not pass validation, activation fails. Call this action to start processing pipeline tasks of a pipeline you've created using the CreatePipeline and PutPipelineDefinition actions. A pipeline cannot be modified after it has been successfully activated. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.ActivatePipeline Content-Length: 39 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE"} HTTP/1.1 200 x-amzn-RequestId: ee19d5bf-074e-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 2 Date: Mon, 12 Nov 2012 17:50:53 GMT {}.

See: ActivatePipeline

activatePipeline

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
apPipelineId

-> State ActivatePipeline a

-> m ActivatePipelineResponse

activatePipelineCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
apPipelineId

-> State ActivatePipeline a

-> m (Either ServiceEr ActivatePipelineResponse)

CreatePipeline

Creates a new empty pipeline. When this action succeeds, you can then use the PutPipelineDefinition action to populate the pipeline. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.CreatePipeline Content-Length: 91 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"name": "myPipeline", "uniqueId": "123456789", "description": "This is my first pipeline"} HTTP/1.1 200 x-amzn-RequestId: b16911ce-0774-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 40 Date: Mon, 12 Nov 2012 17:50:53 GMT {"pipelineId": "df-06372391ZG65EXAMPLE"}.

See: CreatePipeline

createPipeline

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cpName

-> Text
cpUniqueId

-> State CreatePipeline a

-> m CreatePipelineResponse

createPipelineCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cpName

-> Text
cpUniqueId

-> State CreatePipeline a

-> m (Either ServiceEr CreatePipelineResponse)

DeletePipeline

Permanently deletes a pipeline, its pipeline definition and its run history. You cannot query or restore a deleted pipeline. AWS Data Pipeline will attempt to cancel instances associated with the pipeline that are currently being processed by task runners. Deleting a pipeline cannot be undone. To temporarily pause a pipeline instead of deleting it, call SetStatus with the status set to Pause on individual components. Components that are paused by SetStatus can be resumed. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.DeletePipeline Content-Length: 50 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE"} x-amzn-RequestId: b7a88c81-0754-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 0 Date: Mon, 12 Nov 2012 17:50:53 GMT Unexpected response: 200, OK, undefined.

See: DeletePipeline

deletePipeline

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dpPipelineId

-> State DeletePipeline a

-> m DeletePipelineResponse

deletePipelineCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dpPipelineId

-> State DeletePipeline a

-> m (Either ServiceEr DeletePipelineResponse)

DescribeObjects

Returns the object definitions for a set of objects associated with the pipeline. Object definitions are composed of a set of fields that define the properties of the object. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.DescribeObjects Content-Length: 98 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE", "objectIds": [Schedule], "evaluateExpressions": true} x-amzn-RequestId: 4c18ea5d-0777-11e2-8a14-21bb8a1f50ef Content-Type: application/x-amz-json-1.1 Content-Length: 1488 Date: Mon, 12 Nov 2012 17:50:53 GMT {"hasMoreResults": false, "pipelineObjects": [ {"fields": [ {"key": "startDateTime", "stringValue": "2012-12-12T00:00:00"}, {"key": "parent", "refValue": Default}, {"key": "sphere", "stringValue": COMPONENT}, {"key": "type", "stringValue": Schedule}, {"key": "period", "stringValue": "1 hour"}, {"key": "endDateTime", "stringValue": "2012-12-21T18:00:00"}, {"key": "version", "stringValue": "1"}, {"key": "status", "stringValue": PENDING}, {"key": "pipelineId", "stringValue": "df-06372391ZG65EXAMPLE"} ], "id": Schedule, "name": Schedule} ] }.

See: DescribeObjects

describeObjects

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m))

=> Text
doPipelineId

-> [Text]
doObjectIds

-> State DescribeObjects a

-> ResumableSource m DescribeObjectsResponse

describeObjectsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m))

=> Text
doPipelineId

-> [Text]
doObjectIds

-> State DescribeObjects a

-> ResumableSource m (Either ServiceEr DescribeObjectsResponse)

DescribePipelines

Retrieve metadata about one or more pipelines. The information retrieved includes the name of the pipeline, the pipeline identifier, its current state, and the user account that owns the pipeline. Using account credentials, you can retrieve metadata about pipelines that you or your IAM users have created. If you are using an IAM user account, you can retrieve metadata about only those pipelines you have read permission for. To retrieve the full pipeline definition instead of metadata about the pipeline, call the GetPipelineDefinition action. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.DescribePipelines Content-Length: 70 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineIds": ["df-08785951KAKJEXAMPLE"] } x-amzn-RequestId: 02870eb7-0736-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 767 Date: Mon, 12 Nov 2012 17:50:53 GMT {"pipelineDescriptionList": [ {"description": "This is my first pipeline", "fields": [ {"key": "pipelineState", "stringValue": SCHEDULED}, {"key": "description", "stringValue": "This is my first pipeline"}, {"key": "name", "stringValue": "myPipeline"}, {"key": "creationTime", "stringValue": "2012-12-13T01:24:06"}, {"key": "id", "stringValue": "df-0937003356ZJEXAMPLE"}, {"key": "sphere", "stringValue": PIPELINE}, {"key": "version", "stringValue": "1"}, {"key": "userId", "stringValue": "924374875933"}, {"key": "@accountId", "stringValue": "924374875933"}, {"key": "uniqueId", "stringValue": "1234567890"} ], "name": "myPipeline", "pipelineId": "df-0937003356ZJEXAMPLE"} ] }.

See: DescribePipelines

describePipelines

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> [Text]
dp1PipelineIds

-> State DescribePipelines a

-> m DescribePipelinesResponse

describePipelinesCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> [Text]
dp1PipelineIds

-> State DescribePipelines a

-> m (Either ServiceEr DescribePipelinesResponse)

EvaluateExpression

Evaluates a string in the context of a specified object. A task runner can use this action to evaluate SQL queries stored in Amazon S3. POST / HTTP1.1 Content-Type: applicationx-amz-json-1.1 X-Amz-Target: DataPipeline.DescribePipelines Content-Length: 164 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-08785951KAKJEXAMPLE", "objectId": Schedule, "expression": "Transform started at #{startDateTime} and finished at #{endDateTime}"} x-amzn-RequestId: 02870eb7-0736-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 103 Date: Mon, 12 Nov 2012 17:50:53 GMT {"evaluatedExpression": "Transform started at 2012-12-12T00:00:00 and finished at 2012-12-21T18:00:00"}.

See: EvaluateExpression

evaluateExpression

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
eePipelineId

-> Text
eeObjectId

-> Text
eeExpression

-> State EvaluateExpression a

-> m EvaluateExpressionResponse

evaluateExpressionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
eePipelineId

-> Text
eeObjectId

-> Text
eeExpression

-> State EvaluateExpression a

-> m (Either ServiceEr EvaluateExpressionResponse)

GetPipelineDefinition

Returns the definition of the specified pipeline. You can call GetPipelineDefinition to retrieve the pipeline definition you provided using PutPipelineDefinition. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.GetPipelineDefinition Content-Length: 40 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE"} x-amzn-RequestId: e28309e5-0776-11e2-8a14-21bb8a1f50ef Content-Type: application/x-amz-json-1.1 Content-Length: 890 Date: Mon, 12 Nov 2012 17:50:53 GMT {"pipelineObjects": [ {"fields": [ {"key": "workerGroup", "stringValue": "workerGroup"} ], "id": Default, "name": Default}, {"fields": [ {"key": "startDateTime", "stringValue": "2012-09-25T17:00:00"}, {"key": "type", "stringValue": Schedule}, {"key": "period", "stringValue": "1 hour"}, {"key": "endDateTime", "stringValue": "2012-09-25T18:00:00"} ], "id": Schedule, "name": Schedule}, {"fields": [ {"key": "schedule", "refValue": Schedule}, {"key": "command", "stringValue": "echo hello"}, {"key": "parent", "refValue": Default}, {"key": "type", "stringValue": ShellCommandActivity} ], "id": SayHello, "name": SayHello} ] }.

See: GetPipelineDefinition

getPipelineDefinition

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
gpdPipelineId

-> State GetPipelineDefinition a

-> m GetPipelineDefinitionResponse

getPipelineDefinitionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
gpdPipelineId

-> State GetPipelineDefinition a

-> m (Either ServiceEr GetPipelineDefinitionResponse)

ListPipelines

Returns a list of pipeline identifiers for all active pipelines. Identifiers are returned only for pipelines you have permission to access. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.ListPipelines Content-Length: 14 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {} Status: x-amzn-RequestId: b3104dc5-0734-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 39 Date: Mon, 12 Nov 2012 17:50:53 GMT {PipelineIdList: [ {"id": "df-08785951KAKJEXAMPLE", "name": MyPipeline}, {"id": "df-08662578ISYEXAMPLE", "name": MySecondPipeline} ] }.

See: ListPipelines

listPipelines :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m)) => State ListPipelines a -> ResumableSource m ListPipelinesResponse

listPipelinesCatch :: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m)) => State ListPipelines a -> ResumableSource m (Either ServiceEr ListPipelinesResponse)

PollForTask

Task runners call this action to receive a task to perform from AWS Data Pipeline. The task runner specifies which tasks it can perform by setting a value for the workerGroup parameter of the PollForTask call. The task returned by PollForTask may come from any of the pipelines that match the workerGroup value passed in by the task runner and that was launched using the IAM user credentials specified by the task runner. If tasks are ready in the work queue, PollForTask returns a response immediately. If no tasks are available in the queue, PollForTask uses long-polling and holds on to a poll connection for up to a 90 seconds during which time the first newly scheduled task is handed to the task runner. To accomodate this, set the socket timeout in your task runner to 90 seconds. The task runner should not call PollForTask again on the same workerGroup until it receives a response, and this may take up to 90 seconds. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.PollForTask Content-Length: 59 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"workerGroup": MyworkerGroup, "hostname": "example.com"} x-amzn-RequestId: 41c713d2-0775-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 39 Date: Mon, 12 Nov 2012 17:50:53 GMT {"taskObject": {"attemptId": "SayHello_2012-12-12T00:00:00_Attempt=1", "objects": {"SayHello_2012-12-12T00:00:00_Attempt=1": {"fields": [ {"key": "componentParent", "refValue": SayHello}, {"key": "scheduledStartTime", "stringValue": "2012-12-12T00:00:00"}, {"key": "parent", "refValue": SayHello}, {"key": "sphere", "stringValue": ATTEMPT}, {"key": "workerGroup", "stringValue": "workerGroup"}, {"key": "instanceParent", "refValue": "SayHello_2012-12-12T00:00:00"}, {"key": "type", "stringValue": ShellCommandActivity}, {"key": "status", "stringValue": WAITING_FOR_RUNNER}, {"key": "version", "stringValue": "1"}, {"key": "schedule", "refValue": Schedule}, {"key": "actualStartTime", "stringValue": "2012-12-13T01:40:50"}, {"key": "command", "stringValue": "echo hello"}, {"key": "scheduledEndTime", "stringValue": "2012-12-12T01:00:00"}, {"key": "activeInstances", "refValue": "SayHello_2012-12-12T00:00:00"}, {"key": "pipelineId", "stringValue": "df-0937003356ZJEXAMPLE"} ], "id": "SayHello_2012-12-12T00:00:00_Attempt=1", "name": "SayHello_2012-12-12T00:00:00_Attempt=1"} }, "pipelineId": "df-0937003356ZJEXAMPLE", "taskId": "2xaM4wRs5zOsIH+g9U3oVHfAgAlbSqU6XduncB0HhZ3xMnmvfePZPn4dIbYXHyWyRK+cU15MqDHwdrvftx4wv+sNS4w34vJfv7QA9aOoOazW28l1GYSb2ZRR0N0paiQp+d1MhSKo10hOTWOsVK5S5Lnx9Qm6omFgXHyIvZRIvTlrQMpr1xuUrflyGOfbFOGpOLpvPE172MYdqpZKnbSS4TcuqgQKSWV2833fEubI57DPOP7ghWa2TcYeSIv4pdLYG53fTuwfbnbdc98g2LNUQzSVhSnt7BoqyNwht2aQ6bUHg9A80+KVpuXuqmz3m1MXwHFgxjdmuesXNOrrlGpeLCcRWD+aGo0RN1NqhQRzNAig8V4GlaPTQzMsRCljKqvrIyAoP3Tt2XEGsHkkQo12rEX8Z90957XX2qKRwhruwYzqGkSLWjINoLdAxUJdpRXRc5DJTrBd3D5mdzn7kY1l7NEh4kFHJDt3Cx4Z3Mk8MYCACyCkCEyy9DwuPi66cLz0NBcgbCM5LKjTBOwo1m+am+pvM1kSposE9FPP1+RFGb8k6jQBTJx3TRz1yKilnGXQTZ5xvdOFpJrklIT0OXP1MG3+auM9FlJA+1dX90QoNJE5z7axmKMOGXUdkqFe2kiDkorqjxwDvc0Js9pVKfKvAmW8YqUbmI9l0ERpWCXXnLVHNmPWz3jaPY+OBAmuJWDmxBZ8p94aEDg4BVXQ7LvsKQ3DLYhaB7yJ390CJT+i0mm+EBqY60V6YikPSWDFrYQNPi2b1DgE19mX8zHqw8qprIl4yh1Ckx2Iige4EnN5ktOoIxnASxAwTzcE2skxdw5KlHDF+UTj71m16CRdIaKlXijlfNlNzUBobNSadCQn3G5NoO501wPKI:XO50TgDNyo8EXAMPLEg==:1"} }.

See: PollForTask

pollForTask

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
pftWorkerGroup

-> State PollForTask a

-> m PollForTaskResponse

pollForTaskCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
pftWorkerGroup

-> State PollForTask a

-> m (Either ServiceEr PollForTaskResponse)

PutPipelineDefinition

Adds tasks, schedules, and preconditions that control the behavior of the pipeline. You can use PutPipelineDefinition to populate a new pipeline. PutPipelineDefinition also validates the configuration as it adds it to the pipeline. Changes to the pipeline are saved unless one of the following three validation errors exists in the pipeline. An object is missing a name or identifier field. A string or reference field is empty. The number of objects in the pipeline exceeds the maximum allowed objects. Pipeline object definitions are passed to the PutPipelineDefinition action and returned by the GetPipelineDefinition action. Example 1 This example sets an valid pipeline configuration and returns success. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.PutPipelineDefinition Content-Length: 914 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-0937003356ZJEXAMPLE", "pipelineObjects": [ {"id": Default, "name": Default, "fields": [ {"key": "workerGroup", "stringValue": "workerGroup"} ] }, {"id": Schedule, "name": Schedule, "fields": [ {"key": "startDateTime", "stringValue": "2012-12-12T00:00:00"}, {"key": "type", "stringValue": Schedule}, {"key": "period", "stringValue": "1 hour"}, {"key": "endDateTime", "stringValue": "2012-12-21T18:00:00"} ] }, {"id": SayHello, "name": SayHello, "fields": [ {"key": "type", "stringValue": ShellCommandActivity}, {"key": "command", "stringValue": "echo hello"}, {"key": "parent", "refValue": Default}, {"key": "schedule", "refValue": Schedule} ] } ] } HTTP/1.1 200 x-amzn-RequestId: f74afc14-0754-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 18 Date: Mon, 12 Nov 2012 17:50:53 GMT {"errored": false} Example 2 This example sets an invalid pipeline configuration (the value for workerGroup is an empty string) and returns an error message. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.PutPipelineDefinition Content-Length: 903 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE", "pipelineObjects": [ {"id": Default, "name": Default, "fields": [ {"key": "workerGroup", "stringValue": ""} ] }, {"id": Schedule, "name": Schedule, "fields": [ {"key": "startDateTime", "stringValue": "2012-09-25T17:00:00"}, {"key": "type", "stringValue": Schedule}, {"key": "period", "stringValue": "1 hour"}, {"key": "endDateTime", "stringValue": "2012-09-25T18:00:00"} ] }, {"id": SayHello, "name": SayHello, "fields": [ {"key": "type", "stringValue": ShellCommandActivity}, {"key": "command", "stringValue": "echo hello"}, {"key": "parent", "refValue": Default}, {"key": "schedule", "refValue": Schedule} ] } ] } HTTP/1.1 200 x-amzn-RequestId: f74afc14-0754-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 18 Date: Mon, 12 Nov 2012 17:50:53 GMT {"__type": "com.amazon.setl.webservice#InvalidRequestException", "message": "Pipeline definition has errors: Could not save the pipeline definition due to FATAL errors: [com.amazon.setl.webservice.ValidationError@108d7ea9] Please call Validate to validate your pipeline"}.

See: PutPipelineDefinition

putPipelineDefinition

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ppdPipelineId

-> [PipelineObject]
ppdPipelineObjects

-> State PutPipelineDefinition a

-> m PutPipelineDefinitionResponse

putPipelineDefinitionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ppdPipelineId

-> [PipelineObject]
ppdPipelineObjects

-> State PutPipelineDefinition a

-> m (Either ServiceEr PutPipelineDefinitionResponse)

QueryObjects

Queries a pipeline for the names of objects that match a specified set of conditions. The objects returned by QueryObjects are paginated and then filtered by the value you set for query. This means the action may return an empty result set with a value set for marker. If HasMoreResults is set to True, you should continue to call QueryObjects, passing in the returned value for marker, until HasMoreResults returns False. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.QueryObjects Content-Length: 123 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE", "query": {"selectors": [ ] }, "sphere": PO, "marker": "", "limit": 10} x-amzn-RequestId: 14d704c1-0775-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 72 Date: Mon, 12 Nov 2012 17:50:53 GMT {"hasMoreResults": false, "ids": ["@SayHello_1_2012-09-25T17:00:00"] }.

See: QueryObjects

queryObjects

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m))

=> Text
qoPipelineId

-> Text
qoSphere

-> State QueryObjects a

-> ResumableSource m QueryObjectsResponse

queryObjectsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m))

=> Text
qoPipelineId

-> Text
qoSphere

-> State QueryObjects a

-> ResumableSource m (Either ServiceEr QueryObjectsResponse)

ReportTaskProgress

Updates the AWS Data Pipeline service on the progress of the calling task runner. When the task runner is assigned a task, it should call ReportTaskProgress to acknowledge that it has the task within 2 minutes. If the web service does not recieve this acknowledgement within the 2 minute window, it will assign the task in a subsequent PollForTask call. After this initial acknowledgement, the task runner only needs to report progress every 15 minutes to maintain its ownership of the task. You can change this reporting time from 15 minutes by specifying a reportProgressTimeout field in your pipeline. If a task runner does not report its status after 5 minutes, AWS Data Pipeline will assume that the task runner is unable to process the task and will reassign the task in a subsequent response to PollForTask. task runners should call ReportTaskProgress every 60 seconds. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.ReportTaskProgress Content-Length: 832 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"taskId": "aaGgHT4LuH0T0Y0oLrJRjas5qH0d8cDPADxqq3tn+zCWGELkCdV2JprLreXm1oxeP5EFZHFLJ69kjSsLYE0iYHYBYVGBrB+EpYq7ANEEeGJFnSBMRiXZVA+8UJ3OzcInvXeinqBmBaKwii7hnnKbAXjXiNTXyxgydX1KAyg1AxkwBYG4cfPYMZbuEbQJFJvv5C2+GVXz1w94nKYTeUeepwUOFOuRLS6JVtZoYwpF56E+Yfk1IcGpFOvCZ01B4Bkuu7x3J+MDj6kJgZLAgbCJQtI3eiW3kdGmX0p0I2BdY1ZsX6b4UiSvM3OMj6NEHJCJL4E0ZfitnhCoe24Kvjo6C2hFbZq+eiHPgSXBQMSagkr4vS9c0ChzxH2+LNYvec6bY4kymkaZI1dvOzmpa0FcnGf5AjSK4GpsViZujz6zxFv81qBXzjF04M1775rjV1VUdyKaixiAsJiACNezqZqETidp8d24BDPRhGsj6pBCrnelqGFrkgXEXUsJ+xwMifRC8UVwiKekpAvHUywVk7Ku4jHn3i2VoLRP6FXwpUbelu34iiZ9czpXyLtyPKwxa87dlrnRVURwkcVjOt2Mcrcaqe+cbWHvNRhyrPkkdfSF3ac8wfgVbXvLEB2k9mKc67aD9rvdc1PKX09Tk8BKklsMTpZ3TRCd4NzQlJKigMe8Jat9+1tKj4Ole5ZzW6uyTu2s2iFjEV8KXu4MaiRJyNKCdKeGhhZWY37Qk4NBK4Ppgu+C6Y41dpfOh288SLDEVx0UySlqOEdhba7c6BiPp5r3hKj3mk9lFy5OYp1aoGLeeFmjXveTnPdf2gkWqXXg7AUbJ7jEs1F0lKZQg4szep2gcKyAJXgvXLfJJHcha8Lfb/Ee7wYmyOcAaRpDBoFNSbtoVXar46teIrpho+ZDvynUXvU0grHWGOk=:wn3SgymHZM99bEXAMPLE", "fields": [ {"key": "percentComplete", "stringValue": "50"} ] } x-amzn-RequestId: 640bd023-0775-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 18 Date: Mon, 12 Nov 2012 17:50:53 GMT {"canceled": false}.

See: ReportTaskProgress

reportTaskProgress

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
rtpTaskId

-> State ReportTaskProgress a

-> m ReportTaskProgressResponse

reportTaskProgressCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
rtpTaskId

-> State ReportTaskProgress a

-> m (Either ServiceEr ReportTaskProgressResponse)

ReportTaskRunnerHeartbeat

Task runners call ReportTaskRunnerHeartbeat every 15 minutes to indicate that they are operational. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service can use this call to detect when the task runner application has failed and restart a new instance. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.ReportTaskRunnerHeartbeat Content-Length: 84 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"taskrunnerId": "1234567890", "workerGroup": "wg-12345", "hostname": "example.com"} Status: x-amzn-RequestId: b3104dc5-0734-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 20 Date: Mon, 12 Nov 2012 17:50:53 GMT {"terminate": false}.

See: ReportTaskRunnerHeartbeat

reportTaskRunnerHeartbeat

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
rtrhTaskrunnerId

-> State ReportTaskRunnerHeartbeat a

-> m ReportTaskRunnerHeartbeatResponse

reportTaskRunnerHeartbeatCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
rtrhTaskrunnerId

-> State ReportTaskRunnerHeartbeat a

-> m (Either ServiceEr ReportTaskRunnerHeartbeatResponse)

SetStatus

Requests that the status of an array of physical or logical pipeline objects be updated in the pipeline. This update may not occur immediately, but is eventually consistent. The status that can be set depends on the type of object. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.SetStatus Content-Length: 100 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-0634701J7KEXAMPLE", "objectIds": ["o-08600941GHJWMBR9E2"], "status": "pause"} x-amzn-RequestId: e83b8ab7-076a-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 0 Date: Mon, 12 Nov 2012 17:50:53 GMT Unexpected response: 200, OK, undefined.

See: SetStatus

setStatus

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ssPipelineId

-> [Text]
ssObjectIds

-> Text
ssStatus

-> State SetStatus a

-> m SetStatusResponse

setStatusCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ssPipelineId

-> [Text]
ssObjectIds

-> Text
ssStatus

-> State SetStatus a

-> m (Either ServiceEr SetStatusResponse)

SetTaskStatus

Notifies AWS Data Pipeline that a task is completed and provides information about the final status. The task runner calls this action regardless of whether the task was sucessful. The task runner does not need to call SetTaskStatus for tasks that are canceled by the web service during a call to ReportTaskProgress. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.SetTaskStatus Content-Length: 847 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"taskId": "aaGgHT4LuH0T0Y0oLrJRjas5qH0d8cDPADxqq3tn+zCWGELkCdV2JprLreXm1oxeP5EFZHFLJ69kjSsLYE0iYHYBYVGBrB+EpYq7ANEEeGJFnSBMRiXZVA+8UJ3OzcInvXeinqBmBaKwii7hnnKbAXjXiNTXyxgydX1KAyg1AxkwBYG4cfPYMZbuEbQJFJvv5C2+GVXz1w94nKYTeUeepwUOFOuRLS6JVtZoYwpF56E+Yfk1IcGpFOvCZ01B4Bkuu7x3J+MDj6kJgZLAgbCJQtI3eiW3kdGmX0p0I2BdY1ZsX6b4UiSvM3OMj6NEHJCJL4E0ZfitnhCoe24Kvjo6C2hFbZq+eiHPgSXBQMSagkr4vS9c0ChzxH2+LNYvec6bY4kymkaZI1dvOzmpa0FcnGf5AjSK4GpsViZujz6zxFv81qBXzjF04M1775rjV1VUdyKaixiAsJiACNezqZqETidp8d24BDPRhGsj6pBCrnelqGFrkgXEXUsJ+xwMifRC8UVwiKekpAvHUywVk7Ku4jHn3i2VoLRP6FXwpUbelu34iiZ9czpXyLtyPKwxa87dlrnRVURwkcVjOt2Mcrcaqe+cbWHvNRhyrPkkdfSF3ac8wfgVbXvLEB2k9mKc67aD9rvdc1PKX09Tk8BKklsMTpZ3TRCd4NzQlJKigMe8Jat9+1tKj4Ole5ZzW6uyTu2s2iFjEV8KXu4MaiRJyNKCdKeGhhZWY37Qk4NBK4Ppgu+C6Y41dpfOh288SLDEVx0UySlqOEdhba7c6BiPp5r3hKj3mk9lFy5OYp1aoGLeeFmjXveTnPdf2gkWqXXg7AUbJ7jEs1F0lKZQg4szep2gcKyAJXgvXLfJJHcha8Lfb/Ee7wYmyOcAaRpDBoFNSbtoVXar46teIrpho+ZDvynUXvU0grHWGOk=:wn3SgymHZM99bEXAMPLE", "taskStatus": FINISHED} x-amzn-RequestId: 8c8deb53-0788-11e2-af9c-6bc7a6be6qr8 Content-Type: application/x-amz-json-1.1 Content-Length: 0 Date: Mon, 12 Nov 2012 17:50:53 GMT {}.

See: SetTaskStatus

setTaskStatus

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
stsTaskId

-> TaskStatus
stsTaskStatus

-> State SetTaskStatus a

-> m SetTaskStatusResponse

setTaskStatusCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
stsTaskId

-> TaskStatus
stsTaskStatus

-> State SetTaskStatus a

-> m (Either ServiceEr SetTaskStatusResponse)

ValidatePipelineDefinition

Tests the pipeline definition with a set of validation checks to ensure that it is well formed and can run without error. Example 1 This example sets an valid pipeline configuration and returns success. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.ValidatePipelineDefinition Content-Length: 936 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE", "pipelineObjects": [ {"id": Default, "name": Default, "fields": [ {"key": "workerGroup", "stringValue": MyworkerGroup} ] }, {"id": Schedule, "name": Schedule, "fields": [ {"key": "startDateTime", "stringValue": "2012-09-25T17:00:00"}, {"key": "type", "stringValue": Schedule}, {"key": "period", "stringValue": "1 hour"}, {"key": "endDateTime", "stringValue": "2012-09-25T18:00:00"} ] }, {"id": SayHello, "name": SayHello, "fields": [ {"key": "type", "stringValue": ShellCommandActivity}, {"key": "command", "stringValue": "echo hello"}, {"key": "parent", "refValue": Default}, {"key": "schedule", "refValue": Schedule} ] } ] } x-amzn-RequestId: 92c9f347-0776-11e2-8a14-21bb8a1f50ef Content-Type: application/x-amz-json-1.1 Content-Length: 18 Date: Mon, 12 Nov 2012 17:50:53 GMT {"errored": false} Example 2 This example sets an invalid pipeline configuration and returns the associated set of validation errors. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.ValidatePipelineDefinition Content-Length: 903 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE", "pipelineObjects": [ {"id": Default, "name": Default, "fields": [ {"key": "workerGroup", "stringValue": MyworkerGroup} ] }, {"id": Schedule, "name": Schedule, "fields": [ {"key": "startDateTime", "stringValue": "bad-time"}, {"key": "type", "stringValue": Schedule}, {"key": "period", "stringValue": "1 hour"}, {"key": "endDateTime", "stringValue": "2012-09-25T18:00:00"} ] }, {"id": SayHello, "name": SayHello, "fields": [ {"key": "type", "stringValue": ShellCommandActivity}, {"key": "command", "stringValue": "echo hello"}, {"key": "parent", "refValue": Default}, {"key": "schedule", "refValue": Schedule} ] } ] } x-amzn-RequestId: 496a1f5a-0e6a-11e2-a61c-bd6312c92ddd Content-Type: application/x-amz-json-1.1 Content-Length: 278 Date: Mon, 12 Nov 2012 17:50:53 GMT {"errored": true, "validationErrors": [ {"errors": ["INVALID_FIELD_VALUE: startDateTime value must be a literal datetime value."], "id": Schedule} ] }.

See: ValidatePipelineDefinition

validatePipelineDefinition

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
vpdPipelineId

-> [PipelineObject]
vpdPipelineObjects

-> State ValidatePipelineDefinition a

-> m ValidatePipelineDefinitionResponse

validatePipelineDefinitionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
vpdPipelineId

-> [PipelineObject]
vpdPipelineObjects

-> State ValidatePipelineDefinition a

-> m (Either ServiceEr ValidatePipelineDefinitionResponse)

Re-exported

module Network.AWS.DataPipeline.V2012_10_29

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()