Network.AWS.DataPipeline.V2012_10_29.ReportTaskRunnerHeartbeat

Description

Task runners call ReportTaskRunnerHeartbeat every 15 minutes to indicate that they are operational. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service can use this call to detect when the task runner application has failed and restart a new instance. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.ReportTaskRunnerHeartbeat Content-Length: 84 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"taskrunnerId": "1234567890", "workerGroup": "wg-12345", "hostname": "example.com"} Status: x-amzn-RequestId: b3104dc5-0734-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 20 Date: Mon, 12 Nov 2012 17:50:53 GMT {"terminate": false}.

Documentation

reportTaskRunnerHeartbeat

Arguments

:: Text
_rtrhiTaskrunnerId

-> ReportTaskRunnerHeartbeat

Minimum specification for a ReportTaskRunnerHeartbeat request.

data ReportTaskRunnerHeartbeat

Constructors

ReportTaskRunnerHeartbeat

Fields

_rtrhiTaskrunnerId :: Text

The identifier of the task runner. This value should be unique across your AWS account. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service provides a unique identifier when it launches the application. If you have written a custom task runner, you should assign a unique identifier for the task runner.

_rtrhiHostname :: Maybe Text

The public DNS name of the calling task runner.

_rtrhiWorkerGroup :: Maybe Text

Indicates the type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup in the call to ReportTaskRunnerHeartbeat. There are no wildcard values permitted in workerGroup, the string must be an exact, case-sensitive, match.

rtrhiWorkerGroup :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

rtrhiTaskrunnerId :: Lens' ReportTaskRunnerHeartbeat Text

rtrhiHostname :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

data ReportTaskRunnerHeartbeatResponse

Constructors

ReportTaskRunnerHeartbeatResponse

Fields

_rtrhoTerminate :: Bool

Indicates whether the calling task runner should terminate. If True, the task runner that called ReportTaskRunnerHeartbeat should terminate.

rtrhoTerminate :: Iso' ReportTaskRunnerHeartbeatResponse Bool