Network.AWS.DataPipeline.V2012_10_29.SetStatus

Description

Requests that the status of an array of physical or logical pipeline objects be updated in the pipeline. This update may not occur immediately, but is eventually consistent. The status that can be set depends on the type of object. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.SetStatus Content-Length: 100 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-0634701J7KEXAMPLE", "objectIds": ["o-08600941GHJWMBR9E2"], "status": "pause"} x-amzn-RequestId: e83b8ab7-076a-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 0 Date: Mon, 12 Nov 2012 17:50:53 GMT Unexpected response: 200, OK, undefined.

Documentation

data SetStatus

Constructors

SetStatus

Fields

_ssiPipelineId :: Text

Identifies the pipeline that contains the objects.

_ssiObjectIds :: [Text]

Identifies an array of objects. The corresponding objects can be either physical or components, but not a mix of both types.

_ssiStatus :: Text

Specifies the status to be set on all the objects in objectIds. For components, this can be either PAUSE or RESUME. For instances, this can be either CANCEL, RERUN, or MARK_FINISHED.

ssiStatus :: Lens' SetStatus Text

ssiPipelineId :: Lens' SetStatus Text

ssiObjectIds :: Lens' SetStatus [Text]

data SetStatusResponse

Constructors

SetStatusResponse