Network.AWS.DataPipeline.V2012_10_29.SetTaskStatus

Description

Notifies AWS Data Pipeline that a task is completed and provides information about the final status. The task runner calls this action regardless of whether the task was sucessful. The task runner does not need to call SetTaskStatus for tasks that are canceled by the web service during a call to ReportTaskProgress. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.SetTaskStatus Content-Length: 847 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"taskId": "aaGgHT4LuH0T0Y0oLrJRjas5qH0d8cDPADxqq3tn+zCWGELkCdV2JprLreXm1oxeP5EFZHFLJ69kjSsLYE0iYHYBYVGBrB+EpYq7ANEEeGJFnSBMRiXZVA+8UJ3OzcInvXeinqBmBaKwii7hnnKbAXjXiNTXyxgydX1KAyg1AxkwBYG4cfPYMZbuEbQJFJvv5C2+GVXz1w94nKYTeUeepwUOFOuRLS6JVtZoYwpF56E+Yfk1IcGpFOvCZ01B4Bkuu7x3J+MDj6kJgZLAgbCJQtI3eiW3kdGmX0p0I2BdY1ZsX6b4UiSvM3OMj6NEHJCJL4E0ZfitnhCoe24Kvjo6C2hFbZq+eiHPgSXBQMSagkr4vS9c0ChzxH2+LNYvec6bY4kymkaZI1dvOzmpa0FcnGf5AjSK4GpsViZujz6zxFv81qBXzjF04M1775rjV1VUdyKaixiAsJiACNezqZqETidp8d24BDPRhGsj6pBCrnelqGFrkgXEXUsJ+xwMifRC8UVwiKekpAvHUywVk7Ku4jHn3i2VoLRP6FXwpUbelu34iiZ9czpXyLtyPKwxa87dlrnRVURwkcVjOt2Mcrcaqe+cbWHvNRhyrPkkdfSF3ac8wfgVbXvLEB2k9mKc67aD9rvdc1PKX09Tk8BKklsMTpZ3TRCd4NzQlJKigMe8Jat9+1tKj4Ole5ZzW6uyTu2s2iFjEV8KXu4MaiRJyNKCdKeGhhZWY37Qk4NBK4Ppgu+C6Y41dpfOh288SLDEVx0UySlqOEdhba7c6BiPp5r3hKj3mk9lFy5OYp1aoGLeeFmjXveTnPdf2gkWqXXg7AUbJ7jEs1F0lKZQg4szep2gcKyAJXgvXLfJJHcha8Lfb/Ee7wYmyOcAaRpDBoFNSbtoVXar46teIrpho+ZDvynUXvU0grHWGOk=:wn3SgymHZM99bEXAMPLE", "taskStatus": FINISHED} x-amzn-RequestId: 8c8deb53-0788-11e2-af9c-6bc7a6be6qr8 Content-Type: application/x-amz-json-1.1 Content-Length: 0 Date: Mon, 12 Nov 2012 17:50:53 GMT {}.

Request

data SetTaskStatus

The input of the SetTaskStatus action.

Request constructor

mkSetTaskStatus

Arguments

:: Text
stsTaskId

-> TaskStatus
stsTaskStatus

-> SetTaskStatus

Smart constructor for the minimum required parameters to construct a valid SetTaskStatus request.

Request lenses

stsTaskId :: Lens' SetTaskStatus Text

Identifies the task assigned to the task runner. This value is set in the TaskObject that is returned by the PollForTask action.

stsTaskStatus :: Lens' SetTaskStatus TaskStatus

If FINISHED, the task successfully completed. If FAILED the task ended unsuccessfully. The FALSE value is used by preconditions.

stsErrorId :: Lens' SetTaskStatus (Maybe Text)

If an error occurred during the task, this value specifies an id value that represents the error. This value is set on the physical attempt object. It is used to display error information to the user. It should not start with string Service_ which is reserved by the system.

stsErrorMessage :: Lens' SetTaskStatus (Maybe Text)

If an error occurred during the task, this value specifies a text description of the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

stsErrorStackTrace :: Lens' SetTaskStatus (Maybe Text)

If an error occurred during the task, this value specifies the stack trace associated with the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

Response

data SetTaskStatusResponse

The output from the SetTaskStatus action.

Response constructor

mkSetTaskStatusResponse :: SetTaskStatusResponse

Smart constructor for the minimum required parameters to construct a valid SetTaskStatusResponse response.

This constructor is provided for convenience and testing purposes.