Network.AWS.DataPipeline.V2012_10_29.Types

Description

AWS Data Pipeline is a web service that you can use to automate the movement and transformation of data. With AWS Data Pipeline, you can define data-driven workflows, so that tasks can be dependent on the successful completion of previous tasks.

Service

data DataPipeline

Supported version (2012-10-29) of the AWS Data Pipeline service.

Errors

data family Er a

OperatorType

data OperatorType

Constructors

OperatorTypeBetween
BETWEEN

OperatorTypeEq
EQ

OperatorTypeGe
GE

OperatorTypeLe
LE

OperatorTypeRefEq
REF_EQ

TaskStatus

data TaskStatus

Constructors

TaskStatusFailed
FAILED

TaskStatusFalse
FALSE

TaskStatusFinished
FINISHED

Query

data Query

Query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

mkQuery :: Query

Smart constructor for the minimum required fields to construct a valid Query data type to populate a request.

qSelectors :: Lens' Query [Selector]

List of selectors that define the query. An object must satisfy all of the selectors to match the query.

Field

data Field

A key-value pair that describes a property of a pipeline object. The value is specified as either a string value (StringValue) or a reference to another object (RefValue) but not as both.

mkField

Arguments

:: Text
fKey

-> Field

Smart constructor for the minimum required fields to construct a valid Field data type to populate a request.

fKey :: Lens' Field Text

The field identifier.

fStringValue :: Lens' Field (Maybe Text)

The field value, expressed as a String.

fRefValue :: Lens' Field (Maybe Text)

The field value, expressed as the identifier of another object.

InstanceIdentity

data InstanceIdentity

Identity information for the Amazon EC2 instance that is hosting the task runner. You can get this value by calling the URI, http:/169.254.169.254latestmeta-datainstance-id, from the EC2 instance. For more information, go to Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

mkInstanceIdentity :: InstanceIdentity

Smart constructor for the minimum required fields to construct a valid InstanceIdentity data type to populate a request.

iiDocument :: Lens' InstanceIdentity (Maybe Text)

A description of an Amazon EC2 instance that is generated when the instance is launched and exposed to the instance via the instance metadata service in the form of a JSON representation of an object.

iiSignature :: Lens' InstanceIdentity (Maybe Text)

A signature which can be used to verify the accuracy and authenticity of the information provided in the instance identity document.

Operator

data Operator

Contains a logical operation for comparing the value of a field with a specified value.

mkOperator :: Operator

Smart constructor for the minimum required fields to construct a valid Operator data type to populate a request.

oType :: Lens' Operator (Maybe OperatorType)

The logical operation to be performed: equal (EQ), equal reference (REF_EQ), less than or equal (LE), greater than or equal (GE), or between (BETWEEN). Equal reference (REF_EQ) can be used only with reference fields. The other comparison types can be used only with String fields. The comparison types you can use apply only to certain object fields, as detailed below. The comparison operators EQ and REF_EQ act on the following fields: name sphere parent componentParent instanceParent status scheduledStartTime scheduledEndTime actualStartTime actualEndTime The comparison operators GE, LE, and BETWEEN act on the following fields: scheduledStartTime scheduledEndTime actualStartTime actualEndTime Note that fields beginning with the at sign (@) are read-only and set by the web service. When you name fields, you should choose names containing only alpha-numeric values, as symbols may be reserved by AWS Data Pipeline. User-defined fields that you add to a pipeline should prefix their name with the string "my".

oValues :: Lens' Operator [Text]

The value that the actual field value will be compared with.

PipelineDescription

data PipelineDescription

Contains pipeline metadata.

mkPipelineDescription

Arguments

:: Text
pdPipelineId

-> Text
pdName

-> [Field]
pdFields

-> PipelineDescription

Smart constructor for the minimum required fields to construct a valid PipelineDescription data type to populate a request.

pdPipelineId :: Lens' PipelineDescription Text

The pipeline identifier that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

pdName :: Lens' PipelineDescription Text

Name of the pipeline.

pdFields :: Lens' PipelineDescription [Field]

A list of read-only fields that contain metadata about the pipeline: userId, accountId, and @pipelineState.

pdDescription :: Lens' PipelineDescription (Maybe Text)

Description of the pipeline.

PipelineIdName

data PipelineIdName

Contains the name and identifier of a pipeline.

mkPipelineIdName :: PipelineIdName

Smart constructor for the minimum required fields to construct a valid PipelineIdName data type to populate a request.

pinId :: Lens' PipelineIdName (Maybe Text)

Identifier of the pipeline that was assigned by AWS Data Pipeline. This is a string of the form df-297EG78HU43EEXAMPLE.

pinName :: Lens' PipelineIdName (Maybe Text)

Name of the pipeline.

PipelineObject

data PipelineObject

Contains information about a pipeline object. This can be a logical, physical, or physical attempt pipeline object. The complete set of components of a pipeline defines the pipeline.

mkPipelineObject

Arguments

:: Text
poId

-> Text
poName

-> [Field]
poFields

-> PipelineObject

Smart constructor for the minimum required fields to construct a valid PipelineObject data type to populate a request.

poId :: Lens' PipelineObject Text

Identifier of the object.

poName :: Lens' PipelineObject Text

Name of the object.

poFields :: Lens' PipelineObject [Field]

Key-value pairs that define the properties of the object.

Selector

data Selector

A comparision that is used to determine whether a query should return this object.

mkSelector :: Selector

Smart constructor for the minimum required fields to construct a valid Selector data type to populate a request.

sFieldName :: Lens' Selector (Maybe Text)

The name of the field that the operator will be applied to. The field name is the "key" portion of the field definition in the pipeline definition syntax that is used by the AWS Data Pipeline API. If the field is not set on the object, the condition fails.

sOperator :: Lens' Selector (Maybe Operator)

Contains a logical operation for comparing the value of a field with a specified value.

TaskObject

data TaskObject

An instance of PollForTaskResult, which contains an instance of TaskObject. The returned object contains all the information needed to complete the task that is being assigned to the task runner. One of the fields returned in this object is taskId, which contains an identifier for the task being assigned. The calling task runner uses taskId in subsequent calls to ReportTaskProgress and SetTaskStatus.

mkTaskObject :: TaskObject

Smart constructor for the minimum required fields to construct a valid TaskObject data type to populate a request.

toTaskId :: Lens' TaskObject (Maybe Text)

An internal identifier for the task. This ID is passed to the SetTaskStatus and ReportTaskProgress actions.

toPipelineId :: Lens' TaskObject (Maybe Text)

Identifier of the pipeline that provided the task.

toAttemptId :: Lens' TaskObject (Maybe Text)

Identifier of the pipeline task attempt object. AWS Data Pipeline uses this value to track how many times a task is attempted.

toObjects :: Lens' TaskObject (Map Text PipelineObject)

Connection information for the location where the task runner will publish the output of the task.

ValidationError

data ValidationError

Defines a validation error returned by PutPipelineDefinition or ValidatePipelineDefinition. Validation errors prevent pipeline activation. The set of validation errors that can be returned are defined by AWS Data Pipeline.

mkValidationError :: ValidationError

Smart constructor for the minimum required fields to construct a valid ValidationError data type to populate a request.

veId :: Lens' ValidationError (Maybe Text)

The identifier of the object that contains the validation error.

veErrors :: Lens' ValidationError [Text]

A description of the validation error.

ValidationWarning

data ValidationWarning

Defines a validation warning returned by PutPipelineDefinition or ValidatePipelineDefinition. Validation warnings do not prevent pipeline activation. The set of validation warnings that can be returned are defined by AWS Data Pipeline.

mkValidationWarning :: ValidationWarning

Smart constructor for the minimum required fields to construct a valid ValidationWarning data type to populate a request.

vwId :: Lens' ValidationWarning (Maybe Text)

The identifier of the object that contains the validation warning.

vwWarnings :: Lens' ValidationWarning [Text]

A description of the validation warning.