Network.AWS.DataPipeline.V2012_10_29.ValidatePipelineDefinition

Description

Tests the pipeline definition with a set of validation checks to ensure that it is well formed and can run without error. Example 1 This example sets an valid pipeline configuration and returns success. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.ValidatePipelineDefinition Content-Length: 936 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE", "pipelineObjects": [ {"id": Default, "name": Default, "fields": [ {"key": "workerGroup", "stringValue": MyworkerGroup} ] }, {"id": Schedule, "name": Schedule, "fields": [ {"key": "startDateTime", "stringValue": "2012-09-25T17:00:00"}, {"key": "type", "stringValue": Schedule}, {"key": "period", "stringValue": "1 hour"}, {"key": "endDateTime", "stringValue": "2012-09-25T18:00:00"} ] }, {"id": SayHello, "name": SayHello, "fields": [ {"key": "type", "stringValue": ShellCommandActivity}, {"key": "command", "stringValue": "echo hello"}, {"key": "parent", "refValue": Default}, {"key": "schedule", "refValue": Schedule} ] } ] } x-amzn-RequestId: 92c9f347-0776-11e2-8a14-21bb8a1f50ef Content-Type: application/x-amz-json-1.1 Content-Length: 18 Date: Mon, 12 Nov 2012 17:50:53 GMT {"errored": false} Example 2 This example sets an invalid pipeline configuration and returns the associated set of validation errors. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.ValidatePipelineDefinition Content-Length: 903 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE", "pipelineObjects": [ {"id": Default, "name": Default, "fields": [ {"key": "workerGroup", "stringValue": MyworkerGroup} ] }, {"id": Schedule, "name": Schedule, "fields": [ {"key": "startDateTime", "stringValue": "bad-time"}, {"key": "type", "stringValue": Schedule}, {"key": "period", "stringValue": "1 hour"}, {"key": "endDateTime", "stringValue": "2012-09-25T18:00:00"} ] }, {"id": SayHello, "name": SayHello, "fields": [ {"key": "type", "stringValue": ShellCommandActivity}, {"key": "command", "stringValue": "echo hello"}, {"key": "parent", "refValue": Default}, {"key": "schedule", "refValue": Schedule} ] } ] } x-amzn-RequestId: 496a1f5a-0e6a-11e2-a61c-bd6312c92ddd Content-Type: application/x-amz-json-1.1 Content-Length: 278 Date: Mon, 12 Nov 2012 17:50:53 GMT {"errored": true, "validationErrors": [ {"errors": ["INVALID_FIELD_VALUE: startDateTime value must be a literal datetime value."], "id": Schedule} ] }.

Request

data ValidatePipelineDefinition

The input of the ValidatePipelineDefinition action.

Request constructor

mkValidatePipelineDefinition

Arguments

:: Text
vpdPipelineId

-> [PipelineObject]
vpdPipelineObjects

-> ValidatePipelineDefinition

Smart constructor for the minimum required parameters to construct a valid ValidatePipelineDefinition request.

Request lenses

vpdPipelineId :: Lens' ValidatePipelineDefinition Text

Identifies the pipeline whose definition is to be validated.

vpdPipelineObjects :: Lens' ValidatePipelineDefinition [PipelineObject]

A list of objects that define the pipeline changes to validate against the pipeline.

Response

data ValidatePipelineDefinitionResponse

Contains the output from the ValidatePipelineDefinition action.

Response lenses

vpdrValidationErrors :: Lens' ValidatePipelineDefinitionResponse [ValidationError]

Lists the validation errors that were found by ValidatePipelineDefinition.

vpdrValidationWarnings :: Lens' ValidatePipelineDefinitionResponse [ValidationWarning]

Lists the validation warnings that were found by ValidatePipelineDefinition.

vpdrErrored :: Lens' ValidatePipelineDefinitionResponse Bool

If True, there were validation errors.