Network.AWS.DataPipeline.ValidatePipelineDefinition

Description

Validates the specified pipeline definition to ensure that it is well formed and can be run without error.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_ValidatePipelineDefinition.html

Request

data ValidatePipelineDefinition

Request constructor

validatePipelineDefinition

Arguments

:: Text
vpdPipelineId

-> ValidatePipelineDefinition

ValidatePipelineDefinition constructor.

The fields accessible through corresponding lenses are:

vpdParameterObjects :: [ParameterObject]

vpdParameterValues :: [ParameterValue]

vpdPipelineId :: Text

vpdPipelineObjects :: [PipelineObject]

Request lenses

vpdParameterObjects :: Lens' ValidatePipelineDefinition [ParameterObject]

The parameter objects used with the pipeline.

vpdParameterValues :: Lens' ValidatePipelineDefinition [ParameterValue]

The parameter values used with the pipeline.

vpdPipelineId :: Lens' ValidatePipelineDefinition Text

The ID of the pipeline.

vpdPipelineObjects :: Lens' ValidatePipelineDefinition [PipelineObject]

The objects that define the pipeline changes to validate against the pipeline.

Response

data ValidatePipelineDefinitionResponse

Response constructor

validatePipelineDefinitionResponse

Arguments

:: Bool
vpdrErrored

-> ValidatePipelineDefinitionResponse

ValidatePipelineDefinitionResponse constructor.

The fields accessible through corresponding lenses are:

vpdrErrored :: Bool

vpdrValidationErrors :: [ValidationError]

vpdrValidationWarnings :: [ValidationWarning]

Response lenses

vpdrErrored :: Lens' ValidatePipelineDefinitionResponse Bool

Indicates whether there were validation errors.

vpdrValidationErrors :: Lens' ValidatePipelineDefinitionResponse [ValidationError]

Any validation errors that were found.

vpdrValidationWarnings :: Lens' ValidatePipelineDefinitionResponse [ValidationWarning]

Any validation warnings that were found.