Network.AWS.DataPipeline.ActivatePipeline

Description

Validates the specified pipeline and starts processing pipeline tasks. If the pipeline does not pass validation, activation fails.

If you need to pause the pipeline to investigate an issue with a component, such as a data source or script, call DeactivatePipeline.

To activate a finished pipeline, modify the end date for the pipeline and then activate it.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_ActivatePipeline.html

Request

data ActivatePipeline

Request constructor

activatePipeline

Arguments

:: Text
apPipelineId

-> ActivatePipeline

ActivatePipeline constructor.

The fields accessible through corresponding lenses are:

apParameterValues :: [ParameterValue]

apPipelineId :: Text

apStartTimestamp :: Maybe UTCTime

Request lenses

apParameterValues :: Lens' ActivatePipeline [ParameterValue]

A list of parameter values to pass to the pipeline at activation.

apPipelineId :: Lens' ActivatePipeline Text

The ID of the pipeline.

apStartTimestamp :: Lens' ActivatePipeline (Maybe UTCTime)

The date and time to resume the pipeline. By default, the pipeline resumes from the last completed execution.

Response

data ActivatePipelineResponse

Response constructor

activatePipelineResponse :: ActivatePipelineResponse

ActivatePipelineResponse constructor.