Network.AWS.DataPipeline.DeactivatePipeline

Description

Deactivates the specified running pipeline. The pipeline is set to the DEACTIVATING state until the deactivation process completes.

To resume a deactivated pipeline, use ActivatePipeline . By default, the pipeline resumes from the last completed execution. Optionally, you can specify the date and time to resume the pipeline.

Creating a Request

deactivatePipeline

Arguments

:: Text
dPipelineId

-> DeactivatePipeline

Creates a value of DeactivatePipeline with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dCancelActive - Indicates whether to cancel any running objects. The default is true, which sets the state of any running objects to CANCELED . If this value is false, the pipeline is deactivated after all running objects finish.

dPipelineId - The ID of the pipeline.

data DeactivatePipeline

Contains the parameters for DeactivatePipeline.

See: deactivatePipeline smart constructor.

Request Lenses

dCancelActive :: Lens' DeactivatePipeline (Maybe Bool)

Indicates whether to cancel any running objects. The default is true, which sets the state of any running objects to CANCELED . If this value is false, the pipeline is deactivated after all running objects finish.

dPipelineId :: Lens' DeactivatePipeline Text

The ID of the pipeline.

Destructuring the Response

deactivatePipelineResponse

Arguments

:: Int
drsResponseStatus

-> DeactivatePipelineResponse

Creates a value of DeactivatePipelineResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsResponseStatus - -- | The response status code.

data DeactivatePipelineResponse

Contains the output of DeactivatePipeline.

See: deactivatePipelineResponse smart constructor.

Response Lenses

drsResponseStatus :: Lens' DeactivatePipelineResponse Int

- | The response status code.