Network.AWS.DataPipeline.DeletePipeline

Description

Deletes a pipeline, its pipeline definition, and its run history. AWS Data Pipeline attempts to cancel instances associated with the pipeline that are currently being processed by task runners.

Deleting a pipeline cannot be undone. You cannot query or restore a deleted pipeline. To temporarily pause a pipeline instead of deleting it, call SetStatus with the status set to PAUSE on individual components. Components that are paused by SetStatus can be resumed.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_DeletePipeline.html

Request

data DeletePipeline

Request constructor

deletePipeline

Arguments

:: Text
dpPipelineId

-> DeletePipeline

DeletePipeline constructor.

The fields accessible through corresponding lenses are:

dpPipelineId :: Text

Request lenses

dpPipelineId :: Lens' DeletePipeline Text

The ID of the pipeline.

Response

data DeletePipelineResponse

Response constructor

deletePipelineResponse :: DeletePipelineResponse

DeletePipelineResponse constructor.