Network.AWS.DataPipeline.DescribeObjects

Description

Gets the object definitions for a set of objects associated with the pipeline. Object definitions are composed of a set of fields that define the properties of the object.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_DescribeObjects.html

Request

data DescribeObjects

Request constructor

describeObjects

Arguments

:: Text
doPipelineId

-> DescribeObjects

DescribeObjects constructor.

The fields accessible through corresponding lenses are:

doEvaluateExpressions :: Maybe Bool

doMarker :: Maybe Text

doObjectIds :: [Text]

doPipelineId :: Text

Request lenses

doEvaluateExpressions :: Lens' DescribeObjects (Maybe Bool)

Indicates whether any expressions in the object should be evaluated when the object descriptions are returned.

doMarker :: Lens' DescribeObjects (Maybe Text)

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call DescribeObjects with the marker value from the previous call to retrieve the next set of results.

doObjectIds :: Lens' DescribeObjects [Text]

The IDs of the pipeline objects that contain the definitions to be described. You can pass as many as 25 identifiers in a single call to DescribeObjects.

doPipelineId :: Lens' DescribeObjects Text

The ID of the pipeline that contains the object definitions.

Response

data DescribeObjectsResponse

Response constructor

describeObjectsResponse :: DescribeObjectsResponse

DescribeObjectsResponse constructor.

The fields accessible through corresponding lenses are:

dorHasMoreResults :: Maybe Bool

dorMarker :: Maybe Text

dorPipelineObjects :: [PipelineObject]

Response lenses

dorHasMoreResults :: Lens' DescribeObjectsResponse (Maybe Bool)

Indicates whether there are more results to return.

dorMarker :: Lens' DescribeObjectsResponse (Maybe Text)

The starting point for the next page of results. To view the next page of results, call DescribeObjects again with this marker value. If the value is null, there are no more results.

dorPipelineObjects :: Lens' DescribeObjectsResponse [PipelineObject]

An array of object definitions.