Network.AWS.DataPipeline.EvaluateExpression

Description

Task runners call EvaluateExpression to evaluate a string in the context of the specified object. For example, a task runner can evaluate SQL queries stored in Amazon S3.

Creating a Request

evaluateExpression

Arguments

:: Text
eePipelineId

-> Text
eeObjectId

-> Text
eeExpression

-> EvaluateExpression

Creates a value of EvaluateExpression with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eePipelineId - The ID of the pipeline.

eeObjectId - The ID of the object.

eeExpression - The expression to evaluate.

data EvaluateExpression

Contains the parameters for EvaluateExpression.

See: evaluateExpression smart constructor.

Request Lenses

eePipelineId :: Lens' EvaluateExpression Text

The ID of the pipeline.

eeObjectId :: Lens' EvaluateExpression Text

The ID of the object.

eeExpression :: Lens' EvaluateExpression Text

The expression to evaluate.

Destructuring the Response

evaluateExpressionResponse

Arguments

:: Int
eersResponseStatus

-> Text
eersEvaluatedExpression

-> EvaluateExpressionResponse

Creates a value of EvaluateExpressionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eersResponseStatus - -- | The response status code.

eersEvaluatedExpression - The evaluated expression.

data EvaluateExpressionResponse

Contains the output of EvaluateExpression.

See: evaluateExpressionResponse smart constructor.

Response Lenses

eersResponseStatus :: Lens' EvaluateExpressionResponse Int

- | The response status code.

eersEvaluatedExpression :: Lens' EvaluateExpressionResponse Text

The evaluated expression.