Network.AWS.DataPipeline.EvaluateExpression

Description

Evaluates a string in the context of a specified object. A task runner can use this action to evaluate SQL queries stored in Amazon S3. POST / HTTP1.1 Content-Type: applicationx-amz-json-1.1 X-Amz-Target: DataPipeline.DescribePipelines Content-Length: 164 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-08785951KAKJEXAMPLE", "objectId": Schedule, "expression": "Transform started at #{startDateTime} and finished at #{endDateTime}"} x-amzn-RequestId: 02870eb7-0736-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 103 Date: Mon, 12 Nov 2012 17:50:53 GMT {"evaluatedExpression": "Transform started at 2012-12-12T00:00:00 and finished at 2012-12-21T18:00:00"}.

Request

data EvaluateExpression

The input for the EvaluateExpression action.

Request constructor

evaluateExpression

Arguments

:: Text
eePipelineId

-> Text
eeObjectId

-> Text
eeExpression

-> EvaluateExpression

Smart constructor for the minimum required parameters to construct a valid EvaluateExpression request.

The fields accessible through corresponding lenses are:

PipelineId :: Text

ObjectId :: Text

Expression :: Text

Request lenses

eePipelineId :: Lens' EvaluateExpression Text

The identifier of the pipeline.

eeObjectId :: Lens' EvaluateExpression Text

The identifier of the object.

eeExpression :: Lens' EvaluateExpression Text

The expression to evaluate.

Response

data EvaluateExpressionResponse

Contains the output from the EvaluateExpression action.

Response constructor

evaluateExpressionResponse

Arguments

:: Text
eerEvaluatedExpression

-> EvaluateExpressionResponse

Smart constructor for the minimum required parameters to construct a valid EvaluateExpressionResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

EvaluatedExpression :: Text

Response lenses

eerEvaluatedExpression :: Lens' EvaluateExpressionResponse Text

The evaluated expression.