Network.AWS.DataPipeline.GetPipelineDefinition

Description

Gets the definition of the specified pipeline. You can call GetPipelineDefinition to retrieve the pipeline definition that you provided using PutPipelineDefinition .

Creating a Request

getPipelineDefinition

Arguments

:: Text
gpdPipelineId

-> GetPipelineDefinition

Creates a value of GetPipelineDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpdVersion - The version of the pipeline definition to retrieve. Set this parameter to latest (default) to use the last definition saved to the pipeline or active to use the last definition that was activated.

gpdPipelineId - The ID of the pipeline.

data GetPipelineDefinition

Contains the parameters for GetPipelineDefinition.

See: getPipelineDefinition smart constructor.

Request Lenses

gpdVersion :: Lens' GetPipelineDefinition (Maybe Text)

The version of the pipeline definition to retrieve. Set this parameter to latest (default) to use the last definition saved to the pipeline or active to use the last definition that was activated.

gpdPipelineId :: Lens' GetPipelineDefinition Text

The ID of the pipeline.

Destructuring the Response

getPipelineDefinitionResponse

Arguments

:: Int
gpdrsResponseStatus

-> GetPipelineDefinitionResponse

Creates a value of GetPipelineDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpdrsPipelineObjects - The objects defined in the pipeline.

gpdrsParameterObjects - The parameter objects used in the pipeline definition.

gpdrsParameterValues - The parameter values used in the pipeline definition.

gpdrsResponseStatus - -- | The response status code.

data GetPipelineDefinitionResponse

Contains the output of GetPipelineDefinition.

See: getPipelineDefinitionResponse smart constructor.

Response Lenses

gpdrsPipelineObjects :: Lens' GetPipelineDefinitionResponse [PipelineObject]

The objects defined in the pipeline.

gpdrsParameterObjects :: Lens' GetPipelineDefinitionResponse [ParameterObject]

The parameter objects used in the pipeline definition.

gpdrsParameterValues :: Lens' GetPipelineDefinitionResponse [ParameterValue]

The parameter values used in the pipeline definition.

gpdrsResponseStatus :: Lens' GetPipelineDefinitionResponse Int

- | The response status code.