Network.AWS.DataPipeline.GetPipelineDefinition

Description

Gets the definition of the specified pipeline. You can call GetPipelineDefinition to retrieve the pipeline definition that you provided using PutPipelineDefinition.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_GetPipelineDefinition.html

Request

data GetPipelineDefinition

Request constructor

getPipelineDefinition

Arguments

:: Text
gpdPipelineId

-> GetPipelineDefinition

GetPipelineDefinition constructor.

The fields accessible through corresponding lenses are:

gpdPipelineId :: Text

gpdVersion :: Maybe Text

Request lenses

gpdPipelineId :: Lens' GetPipelineDefinition Text

The ID of the pipeline.

gpdVersion :: Lens' GetPipelineDefinition (Maybe Text)

The version of the pipeline definition to retrieve. Set this parameter to latest (default) to use the last definition saved to the pipeline or active to use the last definition that was activated.

Response

data GetPipelineDefinitionResponse

Response constructor

getPipelineDefinitionResponse :: GetPipelineDefinitionResponse

GetPipelineDefinitionResponse constructor.

The fields accessible through corresponding lenses are:

gpdrParameterObjects :: [ParameterObject]

gpdrParameterValues :: [ParameterValue]

gpdrPipelineObjects :: [PipelineObject]

Response lenses

gpdrParameterObjects :: Lens' GetPipelineDefinitionResponse [ParameterObject]

The parameter objects used in the pipeline definition.

gpdrParameterValues :: Lens' GetPipelineDefinitionResponse [ParameterValue]

The parameter values used in the pipeline definition.

gpdrPipelineObjects :: Lens' GetPipelineDefinitionResponse [PipelineObject]

The objects defined in the pipeline.