Network.AWS.DataPipeline.ListPipelines

Description

Lists the pipeline identifiers for all active pipelines that you have permission to access.

This operation returns paginated results.

Creating a Request

listPipelines :: ListPipelines

Creates a value of ListPipelines with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpMarker - The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call ListPipelines with the marker value from the previous call to retrieve the next set of results.

data ListPipelines

Contains the parameters for ListPipelines.

See: listPipelines smart constructor.

Request Lenses

lpMarker :: Lens' ListPipelines (Maybe Text)

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call ListPipelines with the marker value from the previous call to retrieve the next set of results.

Destructuring the Response

listPipelinesResponse

Arguments

:: Int
lprsResponseStatus

-> ListPipelinesResponse

Creates a value of ListPipelinesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lprsHasMoreResults - Indicates whether there are more results that can be obtained by a subsequent call.

lprsMarker - The starting point for the next page of results. To view the next page of results, call ListPipelinesOutput again with this marker value. If the value is null, there are no more results.

lprsResponseStatus - -- | The response status code.

lprsPipelineIdList - The pipeline identifiers. If you require additional information about the pipelines, you can use these identifiers to call DescribePipelines and GetPipelineDefinition .

data ListPipelinesResponse

Contains the output of ListPipelines.

See: listPipelinesResponse smart constructor.

Response Lenses

lprsHasMoreResults :: Lens' ListPipelinesResponse (Maybe Bool)

Indicates whether there are more results that can be obtained by a subsequent call.

lprsMarker :: Lens' ListPipelinesResponse (Maybe Text)

The starting point for the next page of results. To view the next page of results, call ListPipelinesOutput again with this marker value. If the value is null, there are no more results.

lprsResponseStatus :: Lens' ListPipelinesResponse Int

- | The response status code.

lprsPipelineIdList :: Lens' ListPipelinesResponse [PipelineIdName]

The pipeline identifiers. If you require additional information about the pipelines, you can use these identifiers to call DescribePipelines and GetPipelineDefinition .