Network.AWS.DataPipeline.ListPipelines

Description

Lists the pipeline identifiers for all active pipelines that you have permission to access.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_ListPipelines.html

Request

data ListPipelines

Request constructor

listPipelines :: ListPipelines

ListPipelines constructor.

The fields accessible through corresponding lenses are:

lpMarker :: Maybe Text

Request lenses

lpMarker :: Lens' ListPipelines (Maybe Text)

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call ListPipelines with the marker value from the previous call to retrieve the next set of results.

Response

data ListPipelinesResponse

Response constructor

listPipelinesResponse :: ListPipelinesResponse

ListPipelinesResponse constructor.

The fields accessible through corresponding lenses are:

lprHasMoreResults :: Maybe Bool

lprMarker :: Maybe Text

lprPipelineIdList :: [PipelineIdName]

Response lenses

lprHasMoreResults :: Lens' ListPipelinesResponse (Maybe Bool)

Indicates whether there are more results that can be obtained by a subsequent call.

lprMarker :: Lens' ListPipelinesResponse (Maybe Text)

The starting point for the next page of results. To view the next page of results, call ListPipelinesOutput again with this marker value. If the value is null, there are no more results.

lprPipelineIdList :: Lens' ListPipelinesResponse [PipelineIdName]

The pipeline identifiers. If you require additional information about the pipelines, you can use these identifiers to call DescribePipelines and GetPipelineDefinition.