Network.AWS.DataPipeline.PollForTask

Description

Task runners call PollForTask to receive a task to perform from AWS Data Pipeline. The task runner specifies which tasks it can perform by setting a value for the workerGroup parameter. The task returned can come from any of the pipelines that match the workerGroup value passed in by the task runner and that was launched using the IAM user credentials specified by the task runner.

If tasks are ready in the work queue, PollForTask returns a response immediately. If no tasks are available in the queue, PollForTask uses long-polling and holds on to a poll connection for up to a 90 seconds, during which time the first newly scheduled task is handed to the task runner. To accomodate this, set the socket timeout in your task runner to 90 seconds. The task runner should not call PollForTask again on the same workerGroup until it receives a response, and this can take up to 90 seconds.

Creating a Request

pollForTask

Arguments

:: Text
pftWorkerGroup

-> PollForTask

Creates a value of PollForTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pftHostname - The public DNS name of the calling task runner.

pftInstanceIdentity - Identity information for the EC2 instance that is hosting the task runner. You can get this value from the instance using http://169.254.169.254/latest/meta-data/instance-id . For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

pftWorkerGroup - The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup in the call to PollForTask . There are no wildcard values permitted in workerGroup ; the string must be an exact, case-sensitive, match.

data PollForTask

Contains the parameters for PollForTask.

See: pollForTask smart constructor.

Request Lenses

pftHostname :: Lens' PollForTask (Maybe Text)

The public DNS name of the calling task runner.

pftInstanceIdentity :: Lens' PollForTask (Maybe InstanceIdentity)

Identity information for the EC2 instance that is hosting the task runner. You can get this value from the instance using http://169.254.169.254/latest/meta-data/instance-id . For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

pftWorkerGroup :: Lens' PollForTask Text

The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup in the call to PollForTask . There are no wildcard values permitted in workerGroup ; the string must be an exact, case-sensitive, match.

Destructuring the Response

pollForTaskResponse

Arguments

:: Int
pftrsResponseStatus

-> PollForTaskResponse

Creates a value of PollForTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pftrsTaskObject - The information needed to complete the task that is being assigned to the task runner. One of the fields returned in this object is taskId , which contains an identifier for the task being assigned. The calling task runner uses taskId in subsequent calls to ReportTaskProgress and SetTaskStatus .

pftrsResponseStatus - -- | The response status code.

data PollForTaskResponse

Contains the output of PollForTask.

See: pollForTaskResponse smart constructor.

Response Lenses

pftrsTaskObject :: Lens' PollForTaskResponse (Maybe TaskObject)

The information needed to complete the task that is being assigned to the task runner. One of the fields returned in this object is taskId , which contains an identifier for the task being assigned. The calling task runner uses taskId in subsequent calls to ReportTaskProgress and SetTaskStatus .

pftrsResponseStatus :: Lens' PollForTaskResponse Int

- | The response status code.