Network.AWS.DataPipeline.PutPipelineDefinition

Description

Adds tasks, schedules, and preconditions to the specified pipeline. You can use PutPipelineDefinition to populate a new pipeline.

PutPipelineDefinition also validates the configuration as it adds it to the pipeline. Changes to the pipeline are saved unless one of the following three validation errors exists in the pipeline.

An object is missing a name or identifier field. * A string or reference field is empty. * The number of objects in the pipeline exceeds the maximum allowed objects. * The pipeline is in a FINISHED state.

Pipeline object definitions are passed to the PutPipelineDefinition action and returned by the GetPipelineDefinition action.

Creating a Request

putPipelineDefinition

Arguments

:: Text
ppdPipelineId

-> PutPipelineDefinition

Creates a value of PutPipelineDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ppdParameterObjects - The parameter objects used with the pipeline.

ppdParameterValues - The parameter values used with the pipeline.

ppdPipelineId - The ID of the pipeline.

ppdPipelineObjects - The objects that define the pipeline. These objects overwrite the existing pipeline definition.

data PutPipelineDefinition

Contains the parameters for PutPipelineDefinition.

See: putPipelineDefinition smart constructor.

Request Lenses

ppdParameterObjects :: Lens' PutPipelineDefinition [ParameterObject]

The parameter objects used with the pipeline.

ppdParameterValues :: Lens' PutPipelineDefinition [ParameterValue]

The parameter values used with the pipeline.

ppdPipelineId :: Lens' PutPipelineDefinition Text

The ID of the pipeline.

ppdPipelineObjects :: Lens' PutPipelineDefinition [PipelineObject]

The objects that define the pipeline. These objects overwrite the existing pipeline definition.

Destructuring the Response

putPipelineDefinitionResponse

Arguments

:: Int
ppdrsResponseStatus

-> Bool
ppdrsErrored

-> PutPipelineDefinitionResponse

Creates a value of PutPipelineDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ppdrsValidationErrors - The validation errors that are associated with the objects defined in pipelineObjects .

ppdrsValidationWarnings - The validation warnings that are associated with the objects defined in pipelineObjects .

ppdrsResponseStatus - -- | The response status code.

ppdrsErrored - Indicates whether there were validation errors, and the pipeline definition is stored but cannot be activated until you correct the pipeline and call PutPipelineDefinition to commit the corrected pipeline.

data PutPipelineDefinitionResponse

Contains the output of PutPipelineDefinition.

See: putPipelineDefinitionResponse smart constructor.

Response Lenses

ppdrsValidationErrors :: Lens' PutPipelineDefinitionResponse [ValidationError]

The validation errors that are associated with the objects defined in pipelineObjects .

ppdrsValidationWarnings :: Lens' PutPipelineDefinitionResponse [ValidationWarning]

The validation warnings that are associated with the objects defined in pipelineObjects .

ppdrsResponseStatus :: Lens' PutPipelineDefinitionResponse Int

- | The response status code.

ppdrsErrored :: Lens' PutPipelineDefinitionResponse Bool

Indicates whether there were validation errors, and the pipeline definition is stored but cannot be activated until you correct the pipeline and call PutPipelineDefinition to commit the corrected pipeline.