Network.AWS.DataPipeline.PutPipelineDefinition

Description

Adds tasks, schedules, and preconditions to the specified pipeline. You can use PutPipelineDefinition to populate a new pipeline.

PutPipelineDefinition also validates the configuration as it adds it to the pipeline. Changes to the pipeline are saved unless one of the following three validation errors exists in the pipeline.

An object is missing a name or identifier field. A string or reference field is empty. The number of objects in the pipeline exceeds the maximum allowed objects. The pipeline is in a FINISHED state. Pipeline object definitions are passed to the PutPipelineDefinition action and returned by the GetPipelineDefinition action.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_PutPipelineDefinition.html

Request

data PutPipelineDefinition

Request constructor

putPipelineDefinition

Arguments

:: Text
ppdPipelineId

-> PutPipelineDefinition

PutPipelineDefinition constructor.

The fields accessible through corresponding lenses are:

ppdParameterObjects :: [ParameterObject]

ppdParameterValues :: [ParameterValue]

ppdPipelineId :: Text

ppdPipelineObjects :: [PipelineObject]

Request lenses

ppdParameterObjects :: Lens' PutPipelineDefinition [ParameterObject]

The parameter objects used with the pipeline.

ppdParameterValues :: Lens' PutPipelineDefinition [ParameterValue]

The parameter values used with the pipeline.

ppdPipelineId :: Lens' PutPipelineDefinition Text

The ID of the pipeline.

ppdPipelineObjects :: Lens' PutPipelineDefinition [PipelineObject]

The objects that define the pipeline. These objects overwrite the existing pipeline definition.

Response

data PutPipelineDefinitionResponse

Response constructor

putPipelineDefinitionResponse

Arguments

:: Bool
ppdrErrored

-> PutPipelineDefinitionResponse

PutPipelineDefinitionResponse constructor.

The fields accessible through corresponding lenses are:

ppdrErrored :: Bool

ppdrValidationErrors :: [ValidationError]

ppdrValidationWarnings :: [ValidationWarning]

Response lenses

ppdrErrored :: Lens' PutPipelineDefinitionResponse Bool

Indicates whether there were validation errors, and the pipeline definition is stored but cannot be activated until you correct the pipeline and call PutPipelineDefinition to commit the corrected pipeline.

ppdrValidationErrors :: Lens' PutPipelineDefinitionResponse [ValidationError]

The validation errors that are associated with the objects defined in pipelineObjects.

ppdrValidationWarnings :: Lens' PutPipelineDefinitionResponse [ValidationWarning]

The validation warnings that are associated with the objects defined in pipelineObjects.