Network.AWS.DataPipeline.QueryObjects

Description

Queries the specified pipeline for the names of objects that match the specified set of conditions.

This operation returns paginated results.

Creating a Request

queryObjects

Arguments

:: Text
qoPipelineId

-> Text
qoSphere

-> QueryObjects

Creates a value of QueryObjects with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

qoQuery - The query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

qoMarker - The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call QueryObjects with the marker value from the previous call to retrieve the next set of results.

qoLimit - The maximum number of object names that QueryObjects will return in a single call. The default value is 100.

qoPipelineId - The ID of the pipeline.

qoSphere - Indicates whether the query applies to components or instances. The possible values are: COMPONENT , INSTANCE , and ATTEMPT .

data QueryObjects

Contains the parameters for QueryObjects.

See: queryObjects smart constructor.

Request Lenses

qoQuery :: Lens' QueryObjects (Maybe Query)

The query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

qoMarker :: Lens' QueryObjects (Maybe Text)

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call QueryObjects with the marker value from the previous call to retrieve the next set of results.

qoLimit :: Lens' QueryObjects (Maybe Int)

The maximum number of object names that QueryObjects will return in a single call. The default value is 100.

qoPipelineId :: Lens' QueryObjects Text

The ID of the pipeline.

qoSphere :: Lens' QueryObjects Text

Indicates whether the query applies to components or instances. The possible values are: COMPONENT , INSTANCE , and ATTEMPT .

Destructuring the Response

queryObjectsResponse

Arguments

:: Int
qorsResponseStatus

-> QueryObjectsResponse

Creates a value of QueryObjectsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

qorsHasMoreResults - Indicates whether there are more results that can be obtained by a subsequent call.

qorsIds - The identifiers that match the query selectors.

qorsMarker - The starting point for the next page of results. To view the next page of results, call QueryObjects again with this marker value. If the value is null, there are no more results.

qorsResponseStatus - -- | The response status code.

data QueryObjectsResponse

Contains the output of QueryObjects.

See: queryObjectsResponse smart constructor.

Response Lenses

qorsHasMoreResults :: Lens' QueryObjectsResponse (Maybe Bool)

Indicates whether there are more results that can be obtained by a subsequent call.

qorsIds :: Lens' QueryObjectsResponse [Text]

The identifiers that match the query selectors.

qorsMarker :: Lens' QueryObjectsResponse (Maybe Text)

The starting point for the next page of results. To view the next page of results, call QueryObjects again with this marker value. If the value is null, there are no more results.

qorsResponseStatus :: Lens' QueryObjectsResponse Int

- | The response status code.