Network.AWS.DataPipeline.QueryObjects

Description

Queries the specified pipeline for the names of objects that match the specified set of conditions.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_QueryObjects.html

Request

data QueryObjects

Request constructor

queryObjects

Arguments

:: Text
qoPipelineId

-> Text
qoSphere

-> QueryObjects

QueryObjects constructor.

The fields accessible through corresponding lenses are:

qoLimit :: Maybe Int

qoMarker :: Maybe Text

qoPipelineId :: Text

qoQuery :: Maybe Query

qoSphere :: Text

Request lenses

qoLimit :: Lens' QueryObjects (Maybe Int)

The maximum number of object names that QueryObjects will return in a single call. The default value is 100.

qoMarker :: Lens' QueryObjects (Maybe Text)

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call QueryObjects with the marker value from the previous call to retrieve the next set of results.

qoPipelineId :: Lens' QueryObjects Text

The ID of the pipeline.

qoQuery :: Lens' QueryObjects (Maybe Query)

The query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

qoSphere :: Lens' QueryObjects Text

Indicates whether the query applies to components or instances. The possible values are: COMPONENT, INSTANCE, and ATTEMPT.

Response

data QueryObjectsResponse

Response constructor

queryObjectsResponse :: QueryObjectsResponse

QueryObjectsResponse constructor.

The fields accessible through corresponding lenses are:

qorHasMoreResults :: Maybe Bool

qorIds :: [Text]

qorMarker :: Maybe Text

Response lenses

qorHasMoreResults :: Lens' QueryObjectsResponse (Maybe Bool)

Indicates whether there are more results that can be obtained by a subsequent call.

qorIds :: Lens' QueryObjectsResponse [Text]

The identifiers that match the query selectors.

qorMarker :: Lens' QueryObjectsResponse (Maybe Text)

The starting point for the next page of results. To view the next page of results, call QueryObjects again with this marker value. If the value is null, there are no more results.