Network.AWS.DataPipeline.ReportTaskProgress

Description

Task runners call ReportTaskProgress when assigned a task to acknowledge that it has the task. If the web service does not receive this acknowledgement within 2 minutes, it assigns the task in a subsequent PollForTask call. After this initial acknowledgement, the task runner only needs to report progress every 15 minutes to maintain its ownership of the task. You can change this reporting time from 15 minutes by specifying a reportProgressTimeout field in your pipeline.

If a task runner does not report its status after 5 minutes, AWS Data Pipeline assumes that the task runner is unable to process the task and reassigns the task in a subsequent response to PollForTask . Task runners should call ReportTaskProgress every 60 seconds.

Creating a Request

reportTaskProgress

Arguments

:: Text
rtpTaskId

-> ReportTaskProgress

Creates a value of ReportTaskProgress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtpFields - Key-value pairs that define the properties of the ReportTaskProgressInput object.

rtpTaskId - The ID of the task assigned to the task runner. This value is provided in the response for PollForTask .

data ReportTaskProgress

Contains the parameters for ReportTaskProgress.

See: reportTaskProgress smart constructor.

Request Lenses

rtpFields :: Lens' ReportTaskProgress [Field]

Key-value pairs that define the properties of the ReportTaskProgressInput object.

rtpTaskId :: Lens' ReportTaskProgress Text

The ID of the task assigned to the task runner. This value is provided in the response for PollForTask .

Destructuring the Response

reportTaskProgressResponse

Arguments

:: Int
rtprsResponseStatus

-> Bool
rtprsCanceled

-> ReportTaskProgressResponse

Creates a value of ReportTaskProgressResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtprsResponseStatus - -- | The response status code.

rtprsCanceled - If true, the calling task runner should cancel processing of the task. The task runner does not need to call SetTaskStatus for canceled tasks.

data ReportTaskProgressResponse

Contains the output of ReportTaskProgress.

See: reportTaskProgressResponse smart constructor.

Response Lenses

rtprsResponseStatus :: Lens' ReportTaskProgressResponse Int

- | The response status code.

rtprsCanceled :: Lens' ReportTaskProgressResponse Bool

If true, the calling task runner should cancel processing of the task. The task runner does not need to call SetTaskStatus for canceled tasks.