Network.AWS.DataPipeline.ReportTaskProgress

Description

Task runners call ReportTaskProgress when assigned a task to acknowledge that it has the task. If the web service does not receive this acknowledgement within 2 minutes, it assigns the task in a subsequent PollForTask call. After this initial acknowledgement, the task runner only needs to report progress every 15 minutes to maintain its ownership of the task. You can change this reporting time from 15 minutes by specifying a reportProgressTimeout field in your pipeline.

If a task runner does not report its status after 5 minutes, AWS Data Pipeline assumes that the task runner is unable to process the task and reassigns the task in a subsequent response to PollForTask. Task runners should call ReportTaskProgress every 60 seconds.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_ReportTaskProgress.html

Request

data ReportTaskProgress

Request constructor

reportTaskProgress

Arguments

:: Text
rtpTaskId

-> ReportTaskProgress

ReportTaskProgress constructor.

The fields accessible through corresponding lenses are:

rtpFields :: [Field]

rtpTaskId :: Text

Request lenses

rtpFields :: Lens' ReportTaskProgress [Field]

Key-value pairs that define the properties of the ReportTaskProgressInput object.

rtpTaskId :: Lens' ReportTaskProgress Text

The ID of the task assigned to the task runner. This value is provided in the response for PollForTask.

Response

data ReportTaskProgressResponse

Response constructor

reportTaskProgressResponse

Arguments

:: Bool
rtprCanceled

-> ReportTaskProgressResponse

ReportTaskProgressResponse constructor.

The fields accessible through corresponding lenses are:

rtprCanceled :: Bool

Response lenses

rtprCanceled :: Lens' ReportTaskProgressResponse Bool

If true, the calling task runner should cancel processing of the task. The task runner does not need to call SetTaskStatus for canceled tasks.