Network.AWS.DataPipeline.ReportTaskRunnerHeartbeat

Description

Task runners call ReportTaskRunnerHeartbeat every 15 minutes to indicate that they are operational. If the AWS Data Pipeline Task Runner is launched on a resource managed by AWS Data Pipeline, the web service can use this call to detect when the task runner application has failed and restart a new instance.

Creating a Request

reportTaskRunnerHeartbeat

Arguments

:: Text
rtrhTaskrunnerId

-> ReportTaskRunnerHeartbeat

Creates a value of ReportTaskRunnerHeartbeat with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtrhHostname - The public DNS name of the task runner.

rtrhWorkerGroup - The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup . There are no wildcard values permitted in workerGroup ; the string must be an exact, case-sensitive, match.

rtrhTaskrunnerId - The ID of the task runner. This value should be unique across your AWS account. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service provides a unique identifier when it launches the application. If you have written a custom task runner, you should assign a unique identifier for the task runner.

data ReportTaskRunnerHeartbeat

Contains the parameters for ReportTaskRunnerHeartbeat.

See: reportTaskRunnerHeartbeat smart constructor.

Request Lenses

rtrhHostname :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

The public DNS name of the task runner.

rtrhWorkerGroup :: Lens' ReportTaskRunnerHeartbeat (Maybe Text)

The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup . There are no wildcard values permitted in workerGroup ; the string must be an exact, case-sensitive, match.

rtrhTaskrunnerId :: Lens' ReportTaskRunnerHeartbeat Text

The ID of the task runner. This value should be unique across your AWS account. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service provides a unique identifier when it launches the application. If you have written a custom task runner, you should assign a unique identifier for the task runner.

Destructuring the Response

reportTaskRunnerHeartbeatResponse

Arguments

:: Int
rtrhrsResponseStatus

-> Bool
rtrhrsTerminate

-> ReportTaskRunnerHeartbeatResponse

Creates a value of ReportTaskRunnerHeartbeatResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtrhrsResponseStatus - -- | The response status code.

rtrhrsTerminate - Indicates whether the calling task runner should terminate.

data ReportTaskRunnerHeartbeatResponse

Contains the output of ReportTaskRunnerHeartbeat.

See: reportTaskRunnerHeartbeatResponse smart constructor.

Response Lenses

rtrhrsResponseStatus :: Lens' ReportTaskRunnerHeartbeatResponse Int

- | The response status code.

rtrhrsTerminate :: Lens' ReportTaskRunnerHeartbeatResponse Bool

Indicates whether the calling task runner should terminate.