Network.AWS.DataPipeline.SetStatus

Description

Requests that the status of the specified physical or logical pipeline objects be updated in the specified pipeline. This update might not occur immediately, but is eventually consistent. The status that can be set depends on the type of object (for example, DataNode or Activity). You cannot perform this operation on FINISHED pipelines and attempting to do so returns InvalidRequestException .

Creating a Request

setStatus

Arguments

:: Text
ssPipelineId

-> Text
ssStatus

-> SetStatus

Creates a value of SetStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssPipelineId - The ID of the pipeline that contains the objects.

ssObjectIds - The IDs of the objects. The corresponding objects can be either physical or components, but not a mix of both types.

ssStatus - The status to be set on all the objects specified in objectIds . For components, use PAUSE or RESUME . For instances, use TRY_CANCEL , RERUN , or MARK_FINISHED .

data SetStatus

Contains the parameters for SetStatus.

See: setStatus smart constructor.

Request Lenses

ssPipelineId :: Lens' SetStatus Text

The ID of the pipeline that contains the objects.

ssObjectIds :: Lens' SetStatus [Text]

The IDs of the objects. The corresponding objects can be either physical or components, but not a mix of both types.

ssStatus :: Lens' SetStatus Text

The status to be set on all the objects specified in objectIds . For components, use PAUSE or RESUME . For instances, use TRY_CANCEL , RERUN , or MARK_FINISHED .

Destructuring the Response

setStatusResponse :: SetStatusResponse

Creates a value of SetStatusResponse with the minimum fields required to make a request.

data SetStatusResponse

See: setStatusResponse smart constructor.