Network.AWS.DataPipeline.SetStatus

Description

Requests that the status of the specified physical or logical pipeline objects be updated in the specified pipeline. This update might not occur immediately, but is eventually consistent. The status that can be set depends on the type of object (for example, DataNode or Activity). You cannot perform this operation on FINISHED pipelines and attempting to do so returns InvalidRequestException.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_SetStatus.html

Request

data SetStatus

Request constructor

setStatus

Arguments

:: Text
ssPipelineId

-> Text
ssStatus

-> SetStatus

SetStatus constructor.

The fields accessible through corresponding lenses are:

ssObjectIds :: [Text]

ssPipelineId :: Text

ssStatus :: Text

Request lenses

ssObjectIds :: Lens' SetStatus [Text]

The IDs of the objects. The corresponding objects can be either physical or components, but not a mix of both types.

ssPipelineId :: Lens' SetStatus Text

The ID of the pipeline that contains the objects.

ssStatus :: Lens' SetStatus Text

The status to be set on all the objects specified in objectIds. For components, use PAUSE or RESUME. For instances, use TRY_CANCEL, RERUN, or MARK_FINISHED.

Response

data SetStatusResponse

Response constructor

setStatusResponse :: SetStatusResponse

SetStatusResponse constructor.