Network.AWS.DataPipeline.SetTaskStatus

Description

Task runners call SetTaskStatus to notify AWS Data Pipeline that a task is completed and provide information about the final status. A task runner makes this call regardless of whether the task was sucessful. A task runner does not need to call SetTaskStatus for tasks that are canceled by the web service during a call to ReportTaskProgress.

http://docs.aws.amazon.com/datapipeline/latest/APIReference/API_SetTaskStatus.html

Request

data SetTaskStatus

Request constructor

setTaskStatus

Arguments

:: Text
stsTaskId

-> TaskStatus
stsTaskStatus

-> SetTaskStatus

SetTaskStatus constructor.

The fields accessible through corresponding lenses are:

stsErrorId :: Maybe Text

stsErrorMessage :: Maybe Text

stsErrorStackTrace :: Maybe Text

stsTaskId :: Text

stsTaskStatus :: TaskStatus

Request lenses

stsErrorId :: Lens' SetTaskStatus (Maybe Text)

If an error occurred during the task, this value specifies the error code. This value is set on the physical attempt object. It is used to display error information to the user. It should not start with string Service_ which is reserved by the system.

stsErrorMessage :: Lens' SetTaskStatus (Maybe Text)

If an error occurred during the task, this value specifies a text description of the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

stsErrorStackTrace :: Lens' SetTaskStatus (Maybe Text)

If an error occurred during the task, this value specifies the stack trace associated with the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

stsTaskId :: Lens' SetTaskStatus Text

The ID of the task assigned to the task runner. This value is provided in the response for PollForTask.

stsTaskStatus :: Lens' SetTaskStatus TaskStatus

If FINISHED, the task successfully completed. If FAILED, the task ended unsuccessfully. Preconditions use false.

Response

data SetTaskStatusResponse

Response constructor

setTaskStatusResponse :: SetTaskStatusResponse

SetTaskStatusResponse constructor.