Network.AWS.DataPipeline.Types

Service

data DataPipeline

Version 2012-10-29 of the Amazon Data Pipeline service.

Error

data JSONError :: *

ParameterObject

data ParameterObject

parameterObject

Arguments

:: Text
poId

-> ParameterObject

ParameterObject constructor.

The fields accessible through corresponding lenses are:

poAttributes :: [ParameterAttribute]

poId :: Text

poAttributes :: Lens' ParameterObject [ParameterAttribute]

The attributes of the parameter object.

poId :: Lens' ParameterObject Text

The ID of the parameter object.

PipelineObject

data PipelineObject

pipelineObject

Arguments

:: Text
po1Id

-> Text
po1Name

-> PipelineObject

PipelineObject constructor.

The fields accessible through corresponding lenses are:

po1Fields :: [Field]

po1Id :: Text

po1Name :: Text

po1Fields :: Lens' PipelineObject [Field]

Key-value pairs that define the properties of the object.

po1Id :: Lens' PipelineObject Text

The ID of the object.

po1Name :: Lens' PipelineObject Text

The name of the object.

Tag

data Tag

tag

Arguments

:: Text
tagKey

-> Text
tagValue

-> Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Text

tagValue :: Text

tagKey :: Lens' Tag Text

The key name of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

tagValue :: Lens' Tag Text

The optional value portion of a tag defined by a user. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline DeveloperGuide.

Field

data Field

field

Arguments

:: Text
fKey

-> Field

Field constructor.

The fields accessible through corresponding lenses are:

fKey :: Text

fRefValue :: Maybe Text

fStringValue :: Maybe Text

fKey :: Lens' Field Text

The field identifier.

fRefValue :: Lens' Field (Maybe Text)

The field value, expressed as the identifier of another object.

fStringValue :: Lens' Field (Maybe Text)

The field value, expressed as a String.

ParameterValue

data ParameterValue

parameterValue

Arguments

:: Text
pvId

-> Text
pvStringValue

-> ParameterValue

ParameterValue constructor.

The fields accessible through corresponding lenses are:

pvId :: Text

pvStringValue :: Text

pvId :: Lens' ParameterValue Text

The ID of the parameter value.

pvStringValue :: Lens' ParameterValue Text

The field value, expressed as a String.

Selector

data Selector

selector :: Selector

Selector constructor.

The fields accessible through corresponding lenses are:

sFieldName :: Maybe Text

sOperator :: Maybe Operator

sFieldName :: Lens' Selector (Maybe Text)

The name of the field that the operator will be applied to. The field name is the "key" portion of the field definition in the pipeline definition syntax that is used by the AWS Data Pipeline API. If the field is not set on the object, the condition fails.

sOperator :: Lens' Selector (Maybe Operator)

ParameterAttribute

data ParameterAttribute

parameterAttribute

Arguments

:: Text
paKey

-> Text
paStringValue

-> ParameterAttribute

ParameterAttribute constructor.

The fields accessible through corresponding lenses are:

paKey :: Text

paStringValue :: Text

paKey :: Lens' ParameterAttribute Text

The field identifier.

paStringValue :: Lens' ParameterAttribute Text

The field value, expressed as a String.

Operator

data Operator

operator :: Operator

Operator constructor.

The fields accessible through corresponding lenses are:

oType :: Maybe OperatorType

oValues :: [Text]

oType :: Lens' Operator (Maybe OperatorType)

The logical operation to be performed: equal (EQ), equal reference (REF_EQ), less than or equal (LE), greater than or equal (GE), or between (BETWEEN). Equal reference (REF_EQ) can be used only with reference fields. The other comparison types can be used only with String fields. The comparison types you can use apply only to certain object fields, as detailed below.

The comparison operators EQ and REF_EQ act on the following fields:

name sphere parent componentParent instanceParent status scheduledStartTime scheduledEndTime actualStartTime actualEndTime The comparison operators GE, LE, and BETWEEN act on the following fields:

scheduledStartTime scheduledEndTime actualStartTime actualEndTime Note that fields beginning with the at sign (@) are read-only and set by the web service. When you name fields, you should choose names containing only alpha-numeric values, as symbols may be reserved by AWS Data Pipeline. User-defined fields that you add to a pipeline should prefix their name with the string "my".

oValues :: Lens' Operator [Text]

The value that the actual field value will be compared with.

TaskObject

data TaskObject

taskObject :: TaskObject

TaskObject constructor.

The fields accessible through corresponding lenses are:

toAttemptId :: Maybe Text

toObjects :: HashMap Text PipelineObject

toPipelineId :: Maybe Text

toTaskId :: Maybe Text

toAttemptId :: Lens' TaskObject (Maybe Text)

The ID of the pipeline task attempt object. AWS Data Pipeline uses this value to track how many times a task is attempted.

toObjects :: Lens' TaskObject (HashMap Text PipelineObject)

Connection information for the location where the task runner will publish the output of the task.

toPipelineId :: Lens' TaskObject (Maybe Text)

The ID of the pipeline that provided the task.

toTaskId :: Lens' TaskObject (Maybe Text)

An internal identifier for the task. This ID is passed to the SetTaskStatus and ReportTaskProgress actions.

ValidationError

data ValidationError

validationError :: ValidationError

ValidationError constructor.

The fields accessible through corresponding lenses are:

veErrors :: [Text]

veId :: Maybe Text

veErrors :: Lens' ValidationError [Text]

A description of the validation error.

veId :: Lens' ValidationError (Maybe Text)

The identifier of the object that contains the validation error.

PipelineDescription

data PipelineDescription

pipelineDescription

Arguments

:: Text
pdPipelineId

-> Text
pdName

-> PipelineDescription

PipelineDescription constructor.

The fields accessible through corresponding lenses are:

pdDescription :: Maybe Text

pdFields :: [Field]

pdName :: Text

pdPipelineId :: Text

pdTags :: [Tag]

pdDescription :: Lens' PipelineDescription (Maybe Text)

Description of the pipeline.

pdFields :: Lens' PipelineDescription [Field]

A list of read-only fields that contain metadata about the pipeline: userId, accountId, and @pipelineState.

pdName :: Lens' PipelineDescription Text

The name of the pipeline.

pdPipelineId :: Lens' PipelineDescription Text

The pipeline identifier that was assigned by AWS Data Pipeline. This is a string of the form 'df-297EG78HU43EEXAMPLE'.

pdTags :: Lens' PipelineDescription [Tag]

A list of tags to associated with a pipeline. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

InstanceIdentity

data InstanceIdentity

instanceIdentity :: InstanceIdentity

InstanceIdentity constructor.

The fields accessible through corresponding lenses are:

iiDocument :: Maybe Text

iiSignature :: Maybe Text

iiDocument :: Lens' InstanceIdentity (Maybe Text)

A description of an EC2 instance that is generated when the instance is launched and exposed to the instance via the instance metadata service in the form of a JSON representation of an object.

iiSignature :: Lens' InstanceIdentity (Maybe Text)

A signature which can be used to verify the accuracy and authenticity of the information provided in the instance identity document.

Query

data Query

query :: Query

Query constructor.

The fields accessible through corresponding lenses are:

qSelectors :: [Selector]

qSelectors :: Lens' Query [Selector]

List of selectors that define the query. An object must satisfy all of the selectors to match the query.

OperatorType

data OperatorType

Constructors

OperatorBetween
BETWEEN

OperatorEq
EQ

OperatorGe
GE

OperatorLe
LE

OperatorRefEq
REF_EQ

PipelineIdName

data PipelineIdName

pipelineIdName :: PipelineIdName

PipelineIdName constructor.

The fields accessible through corresponding lenses are:

pinId :: Maybe Text

pinName :: Maybe Text

pinId :: Lens' PipelineIdName (Maybe Text)

The ID of the pipeline that was assigned by AWS Data Pipeline. This is a string of the form 'df-297EG78HU43EEXAMPLE'.

pinName :: Lens' PipelineIdName (Maybe Text)

The name of the pipeline.

TaskStatus

data TaskStatus

Constructors

Failed
FAILED

False'
FALSE

Finished
FINISHED

ValidationWarning

data ValidationWarning

validationWarning :: ValidationWarning

ValidationWarning constructor.

The fields accessible through corresponding lenses are:

vwId :: Maybe Text

vwWarnings :: [Text]

vwId :: Lens' ValidationWarning (Maybe Text)

The identifier of the object that contains the validation warning.

vwWarnings :: Lens' ValidationWarning [Text]

A description of the validation warning.