Network.AWS.DataPipeline.V2012_10_29.GetPipelineDefinition

Description

Returns the definition of the specified pipeline. You can call GetPipelineDefinition to retrieve the pipeline definition you provided using PutPipelineDefinition. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.GetPipelineDefinition Content-Length: 40 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE"} x-amzn-RequestId: e28309e5-0776-11e2-8a14-21bb8a1f50ef Content-Type: application/x-amz-json-1.1 Content-Length: 890 Date: Mon, 12 Nov 2012 17:50:53 GMT {"pipelineObjects": [ {"fields": [ {"key": "workerGroup", "stringValue": "workerGroup"} ], "id": Default, "name": Default}, {"fields": [ {"key": "startDateTime", "stringValue": "2012-09-25T17:00:00"}, {"key": "type", "stringValue": Schedule}, {"key": "period", "stringValue": "1 hour"}, {"key": "endDateTime", "stringValue": "2012-09-25T18:00:00"} ], "id": Schedule, "name": Schedule}, {"fields": [ {"key": "schedule", "refValue": Schedule}, {"key": "command", "stringValue": "echo hello"}, {"key": "parent", "refValue": Default}, {"key": "type", "stringValue": ShellCommandActivity} ], "id": SayHello, "name": SayHello} ] }.

Request

data GetPipelineDefinition

The input for the GetPipelineDefinition action.

Request constructor

mkGetPipelineDefinition

Arguments

:: Text
gpdPipelineId

-> GetPipelineDefinition

Smart constructor for the minimum required parameters to construct a valid GetPipelineDefinition request.

The fields accessible through corresponding lenses are:

PipelineId :: Text

Version :: Maybe Text

Request lenses

gpdPipelineId :: Lens' GetPipelineDefinition Text

The identifier of the pipeline.

gpdVersion :: Lens' GetPipelineDefinition (Maybe Text)

The version of the pipeline definition to retrieve. This parameter accepts the values latest (default) and active. Where latest indicates the last definition saved to the pipeline and active indicates the last definition of the pipeline that was activated.

Response

data GetPipelineDefinitionResponse

Contains the output from the GetPipelineDefinition action.

Response constructor

mkGetPipelineDefinitionResponse :: GetPipelineDefinitionResponse

Smart constructor for the minimum required parameters to construct a valid GetPipelineDefinitionResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

PipelineObjects :: [PipelineObject]

Response lenses

gpdrPipelineObjects :: Lens' GetPipelineDefinitionResponse [PipelineObject]

An array of objects defined in the pipeline.