Network.AWS.DataPipeline.V2012_10_29.QueryObjects

Description

Queries a pipeline for the names of objects that match a specified set of conditions. The objects returned by QueryObjects are paginated and then filtered by the value you set for query. This means the action may return an empty result set with a value set for marker. If HasMoreResults is set to True, you should continue to call QueryObjects, passing in the returned value for marker, until HasMoreResults returns False. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: DataPipeline.QueryObjects Content-Length: 123 Host: datapipeline.us-east-1.amazonaws.com X-Amz-Date: Mon, 12 Nov 2012 17:49:52 GMT Authorization: AuthParams {"pipelineId": "df-06372391ZG65EXAMPLE", "query": {"selectors": [ ] }, "sphere": PO, "marker": "", "limit": 10} x-amzn-RequestId: 14d704c1-0775-11e2-af6f-6bc7a6be60d9 Content-Type: application/x-amz-json-1.1 Content-Length: 72 Date: Mon, 12 Nov 2012 17:50:53 GMT {"hasMoreResults": false, "ids": ["@SayHello_1_2012-09-25T17:00:00"] }.

Request

data QueryObjects

The input for the QueryObjects action.

Request constructor

mkQueryObjects

Arguments

:: Text
qoPipelineId

-> Text
qoSphere

-> QueryObjects

Smart constructor for the minimum required parameters to construct a valid QueryObjects request.

Request lenses

qoPipelineId :: Lens' QueryObjects Text

Identifier of the pipeline to be queried for object names.

qoQuery :: Lens' QueryObjects (Maybe Query)

Query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

qoSphere :: Lens' QueryObjects Text

Specifies whether the query applies to components or instances. Allowable values: COMPONENT, INSTANCE, ATTEMPT.

qoMarker :: Lens' QueryObjects (Maybe Text)

The starting point for the results to be returned. The first time you call QueryObjects, this value should be empty. As long as the action returns HasMoreResults as True, you can call QueryObjects again and pass the marker value from the response to retrieve the next set of results.

qoLimit :: Lens' QueryObjects (Maybe Integer)

Specifies the maximum number of object names that QueryObjects will return in a single call. The default value is 100.

Response

data QueryObjectsResponse

Contains the output from the QueryObjects action.

Response lenses

qorIds :: Lens' QueryObjectsResponse [Text]

A list of identifiers that match the query selectors.

qorMarker :: Lens' QueryObjectsResponse (Maybe Text)

The starting point for the results to be returned. As long as the action returns HasMoreResults as True, you can call QueryObjects again and pass the marker value from the response to retrieve the next set of results.

qorHasMoreResults :: Lens' QueryObjectsResponse Bool

If True, there are more results that can be obtained by a subsequent call to QueryObjects.