Network.AWS.DataPipeline.ValidatePipelineDefinition

Description

Validates the specified pipeline definition to ensure that it is well formed and can be run without error.

Creating a Request

validatePipelineDefinition

Arguments

:: Text
vpdPipelineId

-> ValidatePipelineDefinition

Creates a value of ValidatePipelineDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vpdParameterObjects - The parameter objects used with the pipeline.

vpdParameterValues - The parameter values used with the pipeline.

vpdPipelineId - The ID of the pipeline.

vpdPipelineObjects - The objects that define the pipeline changes to validate against the pipeline.

data ValidatePipelineDefinition

Contains the parameters for ValidatePipelineDefinition.

See: validatePipelineDefinition smart constructor.

Request Lenses

vpdParameterObjects :: Lens' ValidatePipelineDefinition [ParameterObject]

The parameter objects used with the pipeline.

vpdParameterValues :: Lens' ValidatePipelineDefinition [ParameterValue]

The parameter values used with the pipeline.

vpdPipelineId :: Lens' ValidatePipelineDefinition Text

The ID of the pipeline.

vpdPipelineObjects :: Lens' ValidatePipelineDefinition [PipelineObject]

The objects that define the pipeline changes to validate against the pipeline.

Destructuring the Response

validatePipelineDefinitionResponse

Arguments

:: Int
vpdrsResponseStatus

-> Bool
vpdrsErrored

-> ValidatePipelineDefinitionResponse

Creates a value of ValidatePipelineDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vpdrsValidationErrors - Any validation errors that were found.

vpdrsValidationWarnings - Any validation warnings that were found.

vpdrsResponseStatus - -- | The response status code.

vpdrsErrored - Indicates whether there were validation errors.

data ValidatePipelineDefinitionResponse

Contains the output of ValidatePipelineDefinition.

See: validatePipelineDefinitionResponse smart constructor.

Response Lenses

vpdrsValidationErrors :: Lens' ValidatePipelineDefinitionResponse [ValidationError]

Any validation errors that were found.

vpdrsValidationWarnings :: Lens' ValidatePipelineDefinitionResponse [ValidationWarning]

Any validation warnings that were found.

vpdrsResponseStatus :: Lens' ValidatePipelineDefinitionResponse Int

- | The response status code.

vpdrsErrored :: Lens' ValidatePipelineDefinitionResponse Bool

Indicates whether there were validation errors.